% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_epsilon.R
\name{apply_epsilon}
\alias{apply_epsilon}
\title{Custom epsilon-correct method}
\usage{
apply_epsilon(
  fit,
  ADREPORT_name = "Index_ctl",
  eps_name = "eps_Index_ctl",
  data_function = identity,
  inner.control = list(sparse = TRUE, lowrank = TRUE, trace = TRUE)
)
}
\arguments{
\item{fit}{output from \code{\link[FishStatsUtils]{fit_model}}}

\item{ADREPORT_name}{string indicating name of ADREPORT'ed variable}

\item{eps_name}{string indicating name of PARAMETER used internally by TMB
for calculating desired gradient}

\item{data_function}{function to apply to \code{fit$data_list} prior to passing
to \code{\link[TMB]{MakeADFun}}, e.g., for VAST use \code{strip_units}}

\item{inner.control}{List controlling inner optimization.}
}
\value{
Standard output from \code{\link[TMB]{sdreport}}, but with slot
        \code{x$unbiased} added if needed, and adding or replacing values for
        \code{x$unbiased$value} corresponding to \code{ADREPORT_name}
}
\description{
\code{apply_epsilon} uses updates to TMB to implement a faster calculation for epsilon-correction
}
