% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arrowtooth_trends}
\alias{arrowtooth_trends}
\title{Data to demonstrate and test decadal-trends model}
\usage{
data(arrowtooth_trends)
}
\description{
Data sufficient to demonstrate how spatially varying coefficients (SVCs)
     are used to represent decadal trends.  Specifically using public data
     from the Alaska Fisheries Science Center bottom trawl survey
     (processed to add true zeros) for arrowtooth flounder in the eastern Bering Sea,
     accessed Sept. 8, 2022.
}
\details{
\itemize{
  \item sampling_data data-frame of tow-specific biomass
}
}
\author{
James Thorson
}
\keyword{data}
