% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{bilinear_interp}
\alias{bilinear_interp}
\title{Plot vessel effects}
\usage{
bilinear_interp(xyz1, xyz2, xyz3, xypred)
}
\arguments{
\item{xyz1, }{2D location and height of first neighbor}

\item{xyz2, }{2D location and height of second neighbor}

\item{xyz3, }{2D location and height of third neighbor}

\item{xypred, }{2D location of location to interpolate}

\item{TmbData}{Formatted data inputs, from `Data_Fn(...)`}

\item{Sdreport}{TMB output from `TMB::sdreport(Obj)`}

\item{FileName_VYplot}{Full filename (including directory) for center-of-gravity plot}
}
\value{
Return Tagged list of output

Calculate weights for bilinear interpolation

\code{bilinear_interp} calculates three coefficients, used to calculate height via blinear interpolation from height and location of three neighbors

Tagged list of useful output
\describe{
  \item{zpred}{height at location \code{xypred}}
  \item{phi}{Coefficients for interpolation}
}
}
\description{
\code{Vessel_Fn} plots estimated vessel effects for model components
}
