% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Build_TMB_Fn}
\alias{Build_TMB_Fn}
\title{Copy of VAST::make_model}
\usage{
Build_TMB_Fn(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{make_model}}
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?VAST::make_model} to see list of arguments
and usage
}
