% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_cov.R
\name{calc_cov}
\alias{calc_cov}
\title{Calculate covariance}
\usage{
calc_cov(L_z, n_f, n_c, returntype = "covariance_matrix")
}
\arguments{
\item{L_z}{vector of loadings parameters}

\item{n_f}{number of factors (columns of loadings matrix)}

\item{n_c}{number of variables (rows of loadings matrix)}

\item{returntype}{character-string matching output}
}
\description{
\code{calc_cov} calculates loadings matrix and covariance
}

