% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{Calc_Polygon_Areas_and_Polygons_Fn}
\alias{Calc_Polygon_Areas_and_Polygons_Fn}
\title{Copy of FishStatsUtils::calculate_knot_areas}
\usage{
Calc_Polygon_Areas_and_Polygons_Fn(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{calculate_knot_areas}}
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?FishStatsUtils::calculate_knot_areas} to see list of arguments
and usage
}
