% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{calc_synchrony}
\alias{calc_synchrony}
\title{Calculate synchrony among species or locations}
\usage{
calc_synchrony(Report, Data, yearbounds_zz = matrix(c(1, Data$n_t), nrow = 1))
}
\arguments{
\item{Report}{output report, e.g., from \code{Report <- obj$report()}}

\item{Data}{tagged list of input data}

\item{yearbounds_zz}{matrix with two columns, giving first and last years for defining one or more periods (rows) used to
calculate changes in synchrony over time (only used if \code{Options['Calculate_Synchrony']=1})}
}
\value{
Tagged list containing measures of synchrony
\describe{
  \item{phi_xz}{Synchrony index for each site (x) and each period (row of \code{yearbounds_zz})}
  \item{phi_z}{weighted-average of \code{phi_xz} for each period, weighted by average community-abundance at each site in that period}
}
}
\description{
\code{calc_synchrony} calculates synchrony
}
