% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_version.R
\name{calc_version}
\alias{calc_version}
\title{Determine latest version of VAST}
\usage{
calc_version(version = NULL)
}
\arguments{
\item{version}{The default is \code{NULL}, which will cause the function
to look for the latest version of the \code{.cpp} file to compile.
If a version is supplied, then \code{R} will look for that version name
within the appropriate folder the disk.}
}
\value{
A full file path as a character value supplying the location of
the latest, or supplied, version of the VAST code to compile.
}
\description{
Determine latest version of VAST
}
\author{
Kelli Faye Johnson
}

