% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_knot_areas.R
\name{calculate_knot_areas}
\alias{calculate_knot_areas}
\title{Calculate areas and covariates for a given triangulated mesh}
\usage{
calculate_knot_areas(loc_x, Data_Extrap, Covariates = "none", a_el = NULL)
}
\arguments{
\item{loc_x}{location (eastings and northings in kilometers, UTM) for each sample or knot}

\item{Data_Extrap}{the output from e.g., \code{Extrapolation_List <- Prepare_WCGBTS_Extrapolation_Data_Fn()} using slot \code{Extrapolation_List$Data_Extrap}}

\item{Covariates}{character vector giving names of columns from Data_Extrap that should be used as covariates (default="none", which results in a design matrix with one columns of 1s)}

\item{a_el}{a matrix with \code{nrow(Data_Extrap)} rows and l columns, giving the area within the l-th stratum for the e-th row of Data_Extrap}
}
\value{
Tagged list containing distance metrics
\describe{
  \item{a_xl}{A data frame with areas for each knot and each strattum}
}
}
\description{
\code{calculate_knot_areas} builds outputs for a given triangulated mesh used for approximating spatial variation
}
