% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_proportion.R
\name{calculate_proportion}
\alias{calculate_proportion}
\title{Calculate the compositional-expansion}
\usage{
calculate_proportion(
  fit,
  TmbData,
  Index,
  Expansion_cz = NULL,
  year_labels = NULL,
  years_to_plot = NULL,
  strata_names = NULL,
  category_names = NULL,
  sample_size_method = c("Taylor_series", "sample_based"),
  plot_legend = ifelse(TmbData$n_l > 1, TRUE, FALSE),
  DirName = getwd(),
  PlotName = "Proportion.png",
  PlotName2 = "Average.png",
  interval_width = 1,
  width = 6,
  height = 6,
  xlab = "Category",
  ylab = "Proportion",
  n_samples = 250,
  ...
)
}
\arguments{
\item{TmbData}{Formatted data inputs, from \code{\link[VAST]{make_data}}}

\item{Index}{output from \code{FishStatsUtils::plot_biomass_index}}

\item{Expansion_cz}{matrix specifying how densities are expanded when calculating annual indices, with a row for each category \code{c} and two columns.
       The first column specifies whether to calculate annual index for category \code{c} as the weighted-sum across density estimates as follows:
\describe{
  \item{\code{Expansion[c,1]=0}}{ density is weighted by area ("area-weighted expansion", the default) }
  \item{\code{Expansion[c,1]=1}}{ density is weighted by the expanded value for another category ("abundance weighted expansion") }
  \item{\code{Expansion[c,1]=2}}{ the index is calculated as the weighted average of density weighted by the expanded value for another category ("abundance weighted-average expansion") }
  \item{\code{Expansion[c,1]=3}}{ area-weighted abundance is added to the expanded abundance for a prior category ("area-weighted cumulative total") }
  \item{\code{Expansion[c,1]=4}}{ The fraction across categories is calculated, and then multiplied by another area-weighted index ("abundance weighted proportional expansion") }
}
       The 2nd column is used when \code{Expansion[c1,1]=1} or \code{Expansion[c1,1]=2} or \code{Expansion[c1,1]=3} or \code{Expansion[c1,1]=4},
       and specifies the category to use for abundance-weighted expansion/average/summation/proportions,
       where \code{Expansion[c1,2]=c2} and \code{c2} must be lower than \code{c1}.}

\item{year_labels}{character vector specifying names for labeling times \code{t_i}}

\item{years_to_plot}{integer vector, specifying positions of \code{year_labels} for plotting (used to avoid plotting years with no data, etc.)}

\item{strata_names}{names for spatial strata}

\item{category_names}{names for categories (if using package \code{`VAST`})}

\item{sample_size_method}{Method used to calculate the variance in proportions, which is then converted to an approximately equivalent multinomial sample size that can be used as input-sample-size in a subsequent stock assessment model.  Options are:
\describe{
  \item{\code{sample_size_method="Taylor_series"}}{a Taylor-series approximation to the ratio of X/(X+Y) where X is the category-specific index and Y is the index for for all other categories}
  \item{\code{sample_size_method="sample_based"}}{Taking samples from the joint precision of fixed and random effects, calculating proportions for each sample, and then computing the variance across those samples}
}
The sample-based approximation is expected to have higher variance and therefore lower approximate sample size.  However, it may also have poor performance in cases when variance estimates are imprecise (such that the multivariate-normal approximation to joint precision is poor), and has not been thoroughly groundtested in real-world cases.}

\item{plot_legend}{Add legend for labelling colors}

\item{DirName}{Directory for saving plot and table}

\item{PlotName}{Name for plot}

\item{interval_width}{width for confidence intervals}

\item{width}{plot width in inches}

\item{height}{plot height in inches}

\item{...}{list of arguments to pass to \code{plot_index}}
}
\value{
Tagged list of output
\describe{
  \item{Prop_ctl}{Proportion of biomass for each category c, time t, and stratum l}
  \item{Neff_tl}{Effective sample size (median across categories) for each time t and stratum l}
}
}
\description{
\code{calculate_proportion} takes output from a VAST run and calculates the proportion of biomass in different categories
}
\references{
For details regarding multivariate index standardization and expansion see \url{https://cdnsciencepub.com/doi/full/10.1139/cjfas-2018-0015}
}
