% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_extrapolation_info.R
\name{combine_extrapolation_info}
\alias{combine_extrapolation_info}
\title{Combines multiple extrapolation grids}
\usage{
combine_extrapolation_info(..., create_strata_per_region = FALSE)
}
\arguments{
\item{...}{a sequence of outputs from \code{FishStatsUtils::make_extrapolation_info}}
}
\value{
Identical output from \code{FishStatsUtils::make_extrapolation_info}, but combined from each input
}
\description{
\code{combine_extrapolation_info} combines multiple extrapolation grids when combining data from multiple surveysn
}
