% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_lists.R
\name{combine_lists}
\alias{combine_lists}
\title{Combine lists}
\usage{
combine_lists(default, input, args_to_use = NULL, use_partial_matching = FALSE)
}
\arguments{
\item{default}{default values for combined list}

\item{input}{preferred values for combined list}
}
\value{
combined list.
}
\description{
\code{combine_lists} combines two lists with precident given to one over the other
}
