% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{Convert_LL_to_EastNorth_Fn}
\alias{Convert_LL_to_EastNorth_Fn}
\title{Convert from Lat-Long to Eastings-Northings using WGS}
\usage{
Convert_LL_to_EastNorth_Fn(Lon, Lat, crs = NA)
}
\arguments{
\item{Lon}{vector of longitudes}

\item{Lat}{vector of latitudes}

\item{crs}{EPSG reference for coordinate reference system (CRS) defining Eastings-Northings after transformation}
}
\value{
A data frame with the following columns
\describe{
  \item{E_km}{The eastings for each value of Lon (in kilometers)}
  \item{N_km}{The northings for each value of Lat (in kilometers)}
}
}
\description{
\code{Convert_LL_to_EastNorth_Fn} converts from Latitude-Longitude to World Geodetic System Eastings-Northings for a given location
}
