% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{Convert_LL_to_UTM_Fn}
\alias{Convert_LL_to_UTM_Fn}
\title{Convert from Lat-Long to UTM}
\usage{
Convert_LL_to_UTM_Fn(Lon, Lat, zone = NA, flip_around_dateline = FALSE)
}
\arguments{
\item{Lon}{vector of longitudes}

\item{Lat}{vector of latitudes}

\item{zone}{UTM zone (integer between 1 and 60) or alphanumeric CRS code used by package rgdal to convert latitude-longitude coordinates to projection in kilometers; \code{zone=NA} uses UTM and automatically detects the appropriate zone}

\item{flip_around_dateline}{boolean specifying whether to flip Lat-Lon locations around the dateline, and then retransform back (only useful if Lat-Lon straddle the dateline)}
}
\value{
A data frame with the following columns
\describe{
  \item{X}{The UTM eastings for each value of Lon}
  \item{Y}{The UTM northings measured from the equator for each Lat}
}
}
\description{
\code{Convert_LL_to_UTM_Fn} converts from Latitude-Longitude to Universal Transverse Mercator projections for a given location
}
