% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_version_name.R
\name{convert_version_name}
\alias{convert_version_name}
\title{Convert string to semantic version number}
\usage{
convert_version_name(char, split = "_")
}
\arguments{
\item{char}{}
}
\value{
A semantic version numvber
}
\description{
Convert string to semantic version number
}
