% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crossvalidate_Fn.R
\name{Crossvalidate_Fn}
\alias{Crossvalidate_Fn}
\title{K-fold crossvalidation}
\usage{
Crossvalidate_Fn(record_dir, parhat, original_data, group_i = NULL,
  kfold = 10, ...)
}
\arguments{
\item{record_dir, }{a directory with writing privilege where the runs are stored}

\item{parhat, }{a tagged list of parameters from the fitted model, e.g., as generated by \code{parhat <- obj$env$parList()} after covergence}

\item{original_data, }{a tagged list of data for the VAST model}

\item{group_i, }{a vector of positive integers, indicating the k-fold group for each observation (default=NULL, which generates a new group_i with even probability)}

\item{kfold, }{the number of crossvalidation batches used (default=10)}
}
\value{
Everything is stored in record_dir
}
\description{
\code{Crossvalidate_Fn} runs a k-fold crossvalidation analysis on a fitted model object
}

