% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Data_Fn}
\alias{Data_Fn}
\title{Copy of VAST::make_data}
\usage{
Data_Fn(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{make_data}}
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?VAST::make_data} to see list of arguments
and usage
}
