% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{Effect.fit_model}
\alias{Effect.fit_model}
\title{Adapts package \code{effects}}
\usage{
Effect.fit_model(
  focal.predictors,
  mod,
  which_formula = "X1",
  pad_values = c(1),
  category_number = NULL,
  ...
)
}
\arguments{
\item{focal.predictors}{a character vector of one or more predictors in the model in any order.}

\item{mod}{a regression model object.  If no specific method exists for the class of \code{mod}, \code{Effect.default} will be called.}

\item{which_formula}{which formula to use e.g., \code{"X1"}}

\item{category_number}{which category code{c_i} to use when plotting density covariates

If getting the error \code{non-conformable arguments}, consider exploring \code{pad_values}
  The error arises in when constructing the linear predictor without an intercept,
  and the default \code{pad_values = 1} attempts to insert a dummy intercept with estimate and SE
  equal to zero.}

\item{...}{arguments to be passed down.}
}
\description{
Calculate effects for plotting
}
