% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{format_covariates}
\alias{format_covariates}
\title{Format habitat covariate matrix}
\usage{
format_covariates(
  Lat_e,
  Lon_e,
  t_e,
  Cov_ep,
  Extrapolation_List,
  Spatial_List,
  FUN = mean,
  Year_Set = min(t_e):max(t_e),
  na.omit = "error"
)
}
\arguments{
\item{Lat_e, }{Latitude for covariate sample e}

\item{Lon_e, }{Longitude for covariate sample e}

\item{t_e, }{Time (e.g., year) for covariate sample e}

\item{Cov_ep, }{matrix of covariates}

\item{Spatial_List, }{Output from \code{FishStatsUtils::Spatial_Information_Fn}, representing spatial location of knots}

\item{FUN, }{function used to aggregate observations associated with each knot-time combination}

\item{Year_Set, }{Set of times \code{t_e} used when generating \code{Cov_xtp}}

\item{na.omit, }{What to do when some knot-time combination has no observation. Options include \code{"error"} which throw an error, or \code{"time-average"} which fills in the average for other years with observations for each knot}
}
\value{
Tagged list of useful output
\describe{
  \item{Cov_xtp}{3-dimensional array for use in \code{Data_Fn}}
  \item{var_p}{a matrix summarizing the data-level variance, variance among knots, and lost variance when aggregating from data to knots for each covariate}
}
}
\description{
\code{format_covariates} formats an array used by \code{Data_Fn} for habitat (a.k.a. density) covariates
}
