% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{Geostat_Sim}
\alias{Geostat_Sim}
\title{Simulate data}
\usage{
Geostat_Sim(
  Sim_Settings,
  Extrapolation_List,
  Data_Geostat = NULL,
  MakePlot = FALSE,
  DateFile = paste0(getwd(), "/"),
  standardize_fields = FALSE
)
}
\arguments{
\item{Sim_Settings}{an optional tagged list for specifying input parameters (see function code to determine settings)}

\item{Extrapolation_List}{the output from \code{\link{make_extrapolation_info}}}

\item{Data_Geostat}{A data frame with column headers \code{c('Lon','Lat','Year','Vessel','AreaSwept_km2')} containing sample design to mimic}

\item{standardize_fields}{Boolean, whether to ensure that random fields have sample mean and standard deviation equal to their inputted values}
}
\value{
Return Tagged list of output
\describe{
  \item{Data_Geostat}{Simulated data for analysis}
  \item{B_tl}{True biomass for each year and stratum}
}
}
\description{
\code{Geostat_Sim} simulates data for use when testing \code{VAST}
}
\examples{
## Do not run (will be slow, due to simulating fine-scale spatial variation for many sites):
##
## # Prepare inputs
## Database = FishData::download_catch_rates( survey="Eastern_Bering_Sea", species_set="Gadus chalcogrammus", error_tol=0.01 )
## Data_Geostat = data.frame("Lat"=Database[,'Lat'], "Lon"=Database[,'Long'], "Year"=Database[,'Year'], "Vessel"="missing", "AreaSwept_km2"=1 )
## Extrapolation_List = FishStatsUtils::Prepare_Extrapolation_Data_Fn( Region="Eastern_Bering_Sea", strata.limits=data.frame( 'STRATA'="All_areas") )
##
## # Use function
## SimList = FishStatsUtils::Geostat_Sim(Sim_Settings=list(), Extrapolation_List, Data_Geostat=Data_Geostat )
## Data_Sim = SimList$Data_Geostat
}
