% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GOA_pcod_covariate_example}
\alias{GOA_pcod_covariate_example}
\title{Data to demonstrate and test covariate effects}
\usage{
data(GOA_pcod_covariate_example)
}
\description{
Data sufficient to demonstrate spatially varying coefficient models
}
\details{
\itemize{
  \item sampling_data data-frame of biological sampling data and associated covariate measurements
  \item Region region for model demo
  \item strata.limits user-specified stratification of results
  \item X_xtp 3D array of covariates, specified at knots given the use of \code{fine_scale=FALSE}
}
}
\author{
Dave McGowan
}
\keyword{data}
