% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goa_traits}
\alias{goa_traits}
\title{Data to demonstrate and test trait-based multispeces SDM}
\usage{
data(goa_traits)
}
\description{
Data sufficient to demonstrate how spatially varying coefficients (SVCs)
     are used to represent trait effects in a joint (multispecies)
     species distribution model.  Specifically using public data
     from the Alaska Fisheries Science Center bottom trawl survey
     (processed to add true zeros) for numerically abundance fishes in the
     Gulf of Alaska, accessed Sept. 8, 2022.  Also including trait estimates
     from an initial (unpublished) release of FishLife that
     includes morphometric and habitat traits.
}
\details{
\itemize{
  \item sampling_data data-frame of tow-specific biomass
  \item trait_data data-frame of trait-values (for continuous traits) and/or probabilities (for categorical traits)
}
}
\author{
James Thorson
}
\keyword{data}
