% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_parameters.R
\name{list_parameters}
\alias{list_parameters}
\title{List fixed and random effects}
\usage{
list_parameters(Obj, verbose = TRUE)
}
\arguments{
\item{Obj}{Compiled TMB object}
}
\value{
Return Tagged-list of fixed and random effects (returned invisibly)
}
\description{
\code{list_parameters} lists all fixed and random effects
}
