% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_kmeans.R
\name{make_kmeans}
\alias{make_kmeans}
\title{Calculate location for knots approximating spatial variation}
\usage{
make_kmeans(
  n_x,
  loc_orig,
  nstart = 100,
  randomseed = 1,
  iter.max = 1000,
  DirPath = getwd(),
  Save_Results = TRUE,
  kmeans_purpose = "spatial",
  backwards_compatible_kmeans = FALSE
)
}
\arguments{
\item{n_x}{the number of knots to select}

\item{loc_orig}{a matrix with two columns where each row gives the 2-dimensional coordinates to be approximated}

\item{nstart}{the number of times that the k-means algorithm is run while searching for the best solution (default=100)}

\item{randomseed}{a random number seed}

\item{iter.max}{the number of iterations used per k-means algorithm (default=1000)}

\item{DirPath}{a directory where the function looks for a previously-saved output (default is working directory)}

\item{Save_Results}{a boolean stating whether to save the output (Default=TRUE)}

\item{kmeans_purpose}{a character representing whether the call is to calculate "extrapolation" or "spatial" information}

\item{backwards_compatible_kmeans}{a boolean stating how to deal with changes in the kmeans algorithm implemented in R version 3.6.0,
where \code{backwards_compatible_kmeans==TRUE} modifies the default algorithm to maintain backwards compatibility, and
where \code{backwards_compatible_kmeans==FALSE} breaks backwards compatibility between R versions prior to and after R 3.6.0.}
}
\value{
Tagged list containing outputs
\describe{
  \item{centers}{a matrix with 2 columns and n_x rows}
  \item{cluster}{A vector with length \code{nrow(loc_orig)} specifying which row of \code{centers} corresponds to each row of loc_orig}
}
}
\description{
\code{make_kmeans} determines the location for a set of knots for approximating spatial variation
}
