% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_map}
\alias{make_map}
\title{Create a tagged list to turn and off and mirror parameters}
\usage{
make_map(
  DataList,
  TmbParams,
  RhoConfig = c(Beta1 = 0, Beta2 = 0, Epsilon1 = 0, Epsilon2 = 0),
  Npool = 0
)
}
\arguments{
\item{DataList}{output from \code{\link{make_data}}}

\item{TmbParams}{output from \code{\link{make_parameters}}}

\item{RhoConfig}{vector of form \code{c("Beta1"=0,"Beta2"=0,"Epsilon1"=0,"Epsilon2"=0)} specifying whether either intercepts (Beta1 and Beta2)
       or spatio-temporal variation (Epsilon1 and Epsilon2) is structured among time intervals, e.g.
       for component \code{Epsilon2} indicated in the 4rd slot:
\describe{
  \item{\code{RhoConfig[4]=0}}{Each year as fixed effect}
  \item{\code{RhoConfig[4]=1}}{Each year as an independent and identically distributed random effect, thus estimating the variance as fixed effect}
  \item{\code{RhoConfig[4]=2}}{Each year as a random effect following a random walk, thus estimating the variance as fixed effect}
  \item{\code{RhoConfig[4]=3}}{Constant among years as fixed effect}
  \item{\code{RhoConfig[4]=4}}{Each year as a random effect following a first-order autoregressive process, thus estimating the variance as fixed effects and a single first-order autoregression parameter}
  \item{\code{RhoConfig[4]=5}}{Each year as a random effect following a first-order autoregressive process, estimating the variance as fixed effects and a separate first-order autoregression parameter for each category}
  \item{\code{RhoConfig[4]=6}}{Only possible for \code{Epsilon2} or \code{Beta2}, and specifying that that associated hyperparameters parameters have identical values to the first component \code{Epsilon1} or \code{Beta1}}
}
If missing, the default is to assume a value of zero for each element (i.e., \code{RhoConfig[1:4]=0})}

\item{Npool}{A user-level interface to pool hyperparameters for multiple categories.
For categories with few encounters, these hyperparameters are poorly informed
leading to converge difficulties.  A value \code{Npool=10} indicates
that any category with fewer than \code{10} encounters across all years
should have hyperparameters mirrored to the same value.}
}
\description{
Create map to turn off and mirror parameters
}
