% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map_info.R
\name{make_map_info}
\alias{make_map_info}
\title{Get information for adequate mapping for regions}
\usage{
make_map_info(
  Region,
  Extrapolation_List,
  spatial_list = NULL,
  NN_Extrap = spatial_list$PolygonList$NN_Extrap,
  fine_scale = spatial_list$fine_scale,
  Include
)
}
\description{
Function provides information to be used when plotting results
}
