% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parameters.R
\name{make_parameters}
\alias{make_parameters}
\title{Calculate parameter inputs for TMB}
\usage{
make_parameters(Version, DataList, RhoConfig = c(Beta1 = 0, Beta2 = 0,
  Epsilon1 = 0, Epsilon2 = 0))
}
\arguments{
\item{Version}{a version number;  If missing, defaults to latest version using \code{FishStatsUtils::get_latest_version(package="VAST")}}

\item{DataList}{list outputted from \code{make_data}}

\item{RhoConfig}{vector of form c("Beta1"=0,"Beta2"=0,"Epsilon1"=0,"Epsilon2"=0) specifying whether either intercepts (Beta1 and Beta2) or spatio-temporal variation (Epsilon1 and Epsilon2) is structured among time intervals (0: each year as fixed effect; 1: each year as random following IID distribution; 2: each year as random following a random walk; 3: constant among years as fixed effect; 4: each year as random following AR1 process);  If missing, assumed to be zero for each element}
}
\value{
Tagged list containing starting values for all fixed effects (parameters) and random effects (coefficients)
\describe{
  \item{ln_H_input}{two parameters governing geometric anisotropy (rotation matrix H)}
  \item{beta1_ct}{intercepts for encounter probability}
  \item{gamma1_j}{effect of density covariates that are static over time on encounter prob}
  \item{gamma1_ctp}{effect of density covariates that change over time on encounter prob}
  \item{lambda1_k}{effect of catchability covariates on encounter prob}
  \item{L1_z}{trimmed cholesky (i.e., parameters for square-root of covariance) for overdispersion in encounter prob}
  \item{L_omega1_z}{trimmed cholesky pointwise variance in spatial variation in encounter prob}
  \item{L_epsilon1_z}{trimmed cholesky pointwise variance in spatio-temporal variation in encounter prob}
  \item{logkappa1}{governs decorrelation distance in encounter prob}
  \item{Beta_mean1}{average intercept of encounter prob (used with RhoConfig options)}
  \item{logsigmaB1}{SD for intercept of encounter prob (used with RhoConfig options)}
  \item{Beta_rho1}{first-order autoregressive coefficient for intercept of encounter prob (used with RhoConfig options)}
  \item{Epsilon_rho1}{first-order autoregressive coefficient for spatio-temporal variation of encounter prob (used with RhoConfig options)}
  \item{eta1_vf}{overdispersion parameters (i.e., vessel or tow-level effects) on encounter prob}
  \item{Omegainput1_sf}{Spatial variation in encounter prob}
  \item{Epsiloninput1_sft}{Spatio-temporal variation in encounter prob}
  \item{beta2_ct}{intercepts for positive catch rates}
  \item{gamma2_j}{effect of density covariates that are static over time on positive catch rates}
  \item{gamma2_ctp}{effect of density covariates that change over time on positive catch rates}
  \item{lambda2_k}{effect of catchability covariates on positive catch rates}
  \item{L2_z}{trimmed cholesky (i.e., parameters for square-root of covariance) for overdispersion in positive catch rates}
  \item{L_omega2_z}{trimmed cholesky pointwise variance in spatial variation in positive catch rates}
  \item{L_epsilon2_z}{trimmed cholesky pointwise variance in spatio-temporal variation in positive catch rates}
  \item{logkappa2}{governs decorrelation distance in positive catch rates}
  \item{Beta_mean2}{average intercept of positive catch rates (used with RhoConfig options)}
  \item{logsigmaB2}{SD for intercept of positive catch rates (used with RhoConfig options)}
  \item{Beta_rho2}{first-order autoregressive coefficient for intercept of positive catch rates (used with RhoConfig options)}
  \item{Epsilon_rho2}{first-order autoregressive coefficient for spatio-temporal variation of positive catch rates (used with RhoConfig options)}
  \item{eta2_vf}{overdispersion parameters (i.e., vessel or tow-level effects) on positive catch rates}
  \item{Omegainput2_sf}{Spatial variation in positive catch rates}
  \item{Epsiloninput2_sft}{Spatio-temporal variation in positive catch rates}
  \item{logSigmaM}{variance parameters for positive catch ratesA}
}
}
\description{
\code{make_parameters} generates the \code{parameters} input for \code{TMB::MakeADFun}
}
