% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Param_Fn}
\alias{Param_Fn}
\title{Copy of make_parameters}
\usage{
Param_Fn(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{make_parameters}}
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?make_parameters} to see list of arguments and usage
}
