% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{plot.fit_model}
\alias{plot.fit_model}
\title{Print parameter estimates}
\usage{
\method{plot}{fit_model}(x, what = "results", ...)
}
\arguments{
\item{what}{String specifying what elements of results to plot;  options include `extrapolation_grid`, `spatial_mesh`, and `results`}

\item{...}{Arguments passed to \code{\link{plot_results}}}

\item{fit}{Output from \code{\link{fit_model}}}
}
\description{
Print parameter estimates and standard errors.
}
