% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_extrapolation_info.R
\name{plot.make_extrapolation_info}
\alias{plot.make_extrapolation_info}
\title{Plot extrapolation-grid}
\usage{
\method{plot}{make_extrapolation_info}(x, cex = 0.01, land_color = "grey", map_resolution = "medium", ...)
}
\arguments{
\item{x}{Output from \code{\link{make_extrapolation_info}}}

\item{...}{Not used}
}
\description{
Plot extrapolation-grid used for spatial inference
}
