% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spatial_info.R
\name{plot.make_spatial_info}
\alias{plot.make_spatial_info}
\title{Plot spatial information}
\usage{
\method{plot}{make_spatial_info}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{make_spatial_info}}}

\item{...}{Not used}
}
\description{
Plot spatial representation used for model
}
