% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_anisotropy.R
\name{plot_anisotropy}
\alias{plot_anisotropy}
\title{Plot geometric anisotropy}
\usage{
plot_anisotropy(
  Obj,
  FileName,
  ControlList = list(Width = 4, Height = 5, Res = 200, Units = "in"),
  type = "ellipse",
  Report = Obj$report(),
  TmbData = Obj$env$data
)
}
\description{
\code{plot_anisotropy} plots the degree of geometric anisotropy
}
\details{
VAST estimates a matrix \code{H}, representing a linear transformation of coordinates (typically eastings-northings)
when computing a distance measure used to calculate correlations between sites. Geometric anisotropy represents the
tendancy for correlations to decline faster in some direction than others, as represented by \code{H}.
\code{plot_anistropy} visualizes the distance needed to achieve a correlation of approximately 10% from a location
centered at coordinates (0,0).  Therefore, an ellipse with a major (long) axis pointed northwest-southeast will
have correlations that decline slower along this axis then movement northeast-southwest. This ellipse is shown for
both linear predictors, which share the same \code{H} estimate but differ in their overall estimated decorrelation rate.
}
