% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot spatial clusters}
\usage{
plot_clusters(
  fit,
  var_name = "D_gct",
  transform_var = log,
  k = 4,
  method = "ward",
  year_labels = fit$year_labels,
  category_names = fit$category_names,
  map_list = NULL,
  working_dir = getwd(),
  file_name = paste0("Class-", var_name),
  file_name2 = paste0("Class-", var_name, "-averages"),
  replace_Inf_with_NA = TRUE,
  size_threshold = 1e+05,
  col = viridisLite::viridis,
  yaxis_log = TRUE,
  ...
)
}
\arguments{
\item{var_name}{Name of object from \code{fit$Report} that is used.  Only
implemented options are \code{"D_gct","Omega1_gc","Omega2_gc","Epsilon1_gct","Epsilon2_gct"}.}

\item{transform_var}{function to apply to \code{fit$Report[[var_name]]} prior to clustering.
I recommend using log-transform for density, and otherwise using \code{transform_var=identity}.}

\item{k}{an integer scalar or vector with the desired number of groups}

\item{method}{Distance metric. Default \code{method="ward"} is very fast,
but can instead use \code{method="bcdist"} for small problems which calculates
Bray-Curtist dissimilarity using \code{\link[ecodist]{bcdist}} and then applies
Ward clustering}

\item{map_list}{output from \code{\link{make_map_info output}}}

\item{replace_Inf_with_NA}{Boolean whether to replace \code{Inf} or \code{-Inf} values
with \code{NA} prior to clustering, as useful sometimes when \code{var_name="D_gct"},
\code{transform_var=log} and replacing nonencounters with zero.}
}
\description{
\code{plot_clusters} plots the clusters on a map
}
\details{
Code to plot clusters to look for biogeographic shifts, as well as plotting
     averages across all sites-times assigned to a each cluster, for use when
     interpreting each cluster.
}
\references{
For details regarding spatial clustering see \url{https://doi.org/10.1093/icesjms/fsac007}
}
