% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plot_cov}
\alias{plot_cov}
\title{Plot covariance matrix}
\usage{
plot_cov(
  Cov,
  zlim = NULL,
  names = 1:nrow(Cov),
  names2 = names,
  ncolors = 21,
  digits = 2,
  ...
)
}
\arguments{
\item{Cov}{matrix (covariance or correlation) used for plotting}

\item{zlim}{numeric-vector (length 2) defining bounds for color-scale of covariance}

\item{names}{Labels for y-axis}

\item{names2}{Labels for x-axis (on top of covariance)}

\item{ncolors}{Number of colors for color-scale}

\item{digits}{Number of digits for text-labels of covariance}

\item{...}{passed to \code{text} for labelling covariances}
}
\description{
\code{plot_cov} plots and formats a covariance matrix
}
