% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot location of sampling data}
\usage{
plot_data(
  Extrapolation_List,
  Spatial_List,
  Data_Geostat = NULL,
  Lat_i = Data_Geostat[, "Lat"],
  Lon_i = Data_Geostat[, "Lon"],
  Year_i = Data_Geostat[, "Year"],
  PlotDir = getwd(),
  Plot1_name = "Data_and_knots.png",
  Plot2_name = "Data_by_year.png",
  col = "red",
  cex = 0.1,
  pch = 19,
  year_labels,
  projargs = "+proj=longlat",
  map_resolution = "medium",
  land_color = "grey",
  country = NULL,
  ...
)
}
\arguments{
\item{Extrapolation_List}{Output from \code{Prepare_Extrapolation_Data_Fn}}

\item{Spatial_List}{Output from \code{Spatial_Information_Fn}}

\item{Data_Geostat}{data-frame of data (with columns 'E_km', 'N_km', 'Year', 'Lon', 'Lat' at a minimum)}

\item{PlotDir}{Directory for plots}

\item{year_labels}{character vector specifying names for labeling times \code{t_i}}

\item{projargs}{Character passed to \code{\link[sp]{CRS}} and see that documentation for description.
See \url{https://proj.org/operations/projections/index.html} for a list of projections to pass via \code{projargs}.
As a stable default please consider \code{projargs='+proj=natearth +lon_0=0 +units=km'}
where argument \code{+lon_0} allows the user to center eastings on a specified longitude.}

\item{land_color}{color for filling in land (use \code{land_color=rgb(0,0,0,alpha=0)} for transparent land)}

\item{country}{a character vector of country names.}

\item{...}{addition inputs to \code{plot}}
}
\description{
\code{plot_data} produces diagnostics plots for the spatial distribution of data and knots
}
