% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dharma.R
\name{plot_dharma}
\alias{plot_dharma}
\title{Plot DHARMa residuals}
\usage{
plot_dharma(dharmaRes, form = NULL, rank = TRUE, ...)
}
\arguments{
\item{dharmaRes}{Output from \code{\link{plot}} using slot \code{$dharmaRes}}

\item{form}{Optional predictor against which the residuals should be plotted. Default is to used the predicted(simulationOutput).}

\item{rank}{If T, the values provided in form will be rank transformed. This will usually make patterns easier to spot visually, especially if the distribution of the predictor is skewed. If form is a factor, this has no effect.}

\item{...}{Additional arguments to plot / boxplot.}
}
\description{
Plot DHARMa residuals
}
\author{
Zach Oyafuso
}
