% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plot_encounter_diagnostic}
\alias{plot_encounter_diagnostic}
\title{Check predicted encounter probability against observed encounter frequency}
\usage{
plot_encounter_diagnostic(
  Report,
  Data_Geostat,
  cutpoints_z = seq(0, 1, length = 21),
  interval_width = 1.96,
  DirName = paste0(getwd(), "/"),
  PlotName = "Diag--Encounter_prob.png",
  ...
)
}
\arguments{
\item{Report}{tagged list of outputs from TMB model via \code{Obj$report()}}

\item{interval_width}{width for confidence intervals}

\item{DirName}{Directory for saving plot and table}

\item{PlotName}{Name for plot}

\item{...}{arguments passed to \code{par}}
}
\value{
Return Tagged list of output
\describe{
  \item{Diag_i}{Diagnostic output for each sample \code{i}}
  \item{Diag_z}{Diagnostic output for each bin \code{z}}
}
}
\description{
\code{plot_encounter_diagnostic} is a diagnostic function for checking validity of the encounter-probability component of a spatio-temporal model
}
