% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Plot_factors}
\alias{Plot_factors}
\title{Copy of FishStatsUtils::plot_factors}
\usage{
Plot_factors(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{plot_factors}}
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?FishStatsUtils::plot_factors} to see list of arguments and usage
}
