% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index.R
\name{plot_index}
\alias{plot_index}
\title{Plot an index}
\usage{
plot_index(
  Index_ctl,
  sd_Index_ctl = array(0, dim(Index_ctl)),
  year_labels = NULL,
  years_to_plot = NULL,
  strata_names = NULL,
  category_names = NULL,
  scale = "uniform",
  plot_legend = NULL,
  DirName = getwd(),
  PlotName = "Index.png",
  interval_width = 1,
  width = NULL,
  height = NULL,
  xlab = "Year",
  ylab = "Index",
  bounds_type = "whiskers",
  col = NULL,
  col_bounds = NULL,
  Yrange = c(0, NA),
  type = "b",
  plot_lines_args = list(),
  plot_args = list(),
  SampleSize_ctz = NULL,
  Y2range = c(0, NA),
  y2lab = "",
  add = FALSE,
  ...
)
}
\arguments{
\item{Index_ctl}{A matrix or array of time-series estimates for multiple categories \code{c}, years \code{t}, and strata \code{l}}

\item{sd_Index_ctl}{A matrix or array of variances for each estimate}

\item{year_labels}{character vector specifying names for labeling times \code{t_i}}

\item{years_to_plot}{integer vector, specifying positions of \code{year_labels} for plotting (used to avoid plotting years with no data, etc.)}

\item{strata_names}{names for spatial strata}

\item{category_names}{names for categories (if using package \code{`VAST`})}

\item{plot_legend}{Add legend for labelling colors}

\item{DirName}{Directory for saving plot and table}

\item{PlotName}{Name for plot}

\item{interval_width}{width for confidence intervals}

\item{width}{plot width in inches}

\item{height}{plot height in inches}

\item{Yrange}{lower and upper bound for left-hand y-axis, corresponding to input \code{Index_ctl}
(use \code{Yrange[1]=NA} and/or \code{Yrange[2]=NA} for using the lower and upper bound of estimate intervals)}

\item{plot_lines_args}{additional arguments to pass to \code{plot_lines}}

\item{plot_args}{additional arguments to pass to \code{plot}}

\item{SampleSize_ctz}{optional array of sample sizes for each category and year to be plotted on each panel}

\item{Y2range}{lower and upper bound for right-hand y-axis, corresponding to input \code{SampleSize_ctz} (see Yrange for more info)}

\item{...}{list of settings to pass to \code{par} when making plot}
}
\description{
\code{plot_index} takes output from a VAST run and plots a panel figure of time-series estimates
}
