% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_legend.R
\name{plot_legend}
\alias{plot_legend}
\title{Plot standard maps}
\usage{
plot_legend(
  Zlim,
  legend_x = c(0, 0.05),
  legend_y = c(0.05, 0.45),
  cex.legend = 1,
  col = viridisLite::viridis,
  legend_digits = 1,
  f = identity
)
}
\description{
Plot standard maps
}
