% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loadings.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Plotting loadings matrix}
\usage{
plot_loadings(
  L_pj,
  Lsd_pj = NULL,
  At = 1:nrow(L_pj),
  whichfactor = 1,
  addtitle = TRUE,
  LabelPosition = "Right",
  Buffer = c(0, 0.1),
  Labels = rownames(L_pj),
  Cex = 1.2,
  legend_text = "Proportion of explained variance",
  interval_width = 1,
  ...
)
}
\arguments{
\item{L_pj}{Loadings matrix for `p` categories and `j` factors}

\item{whichfactor}{Integer, giving column of the loadings matrix to plot}

\item{addtitle}{Boolean, whether to add a title}

\item{LabelPosition}{Character, where to plot labels for rows of \code{L_pj} (Options: "Right","Above","Below")}

\item{Buffer, }{how much to pad top and bottom of each panel}

\item{Labels, }{Labels for rows of L_pj}
}
\value{
tagged list of outputs
\describe{
  \item{L_pj_rot}{Loadings matrix after rotation}
  \item{Psi_rot}{Factors after rotation}
  \item{Hinv}{Object used for rotation}
}
}
\description{
\code{plot_loadings} plots elements of a column of the loadings matrix
}
