% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.r
\name{plot_maps}
\alias{plot_maps}
\title{Plot standard maps}
\usage{
plot_maps(
  plot_set = 3,
  fit,
  PlotDF,
  projargs = "+proj=longlat",
  Panel = "Category",
  year_labels = NULL,
  years_to_plot = NULL,
  category_names = NULL,
  quiet = FALSE,
  working_dir = getwd(),
  MapSizeRatio = NULL,
  n_cells,
  plot_value = "estimate",
  n_samples = 100,
  country = NULL,
  sample_fixed = TRUE,
  Report = fit$Report,
  TmbData = fit$data_list,
  Obj = fit$tmb_list$Obj,
  extrapolation_list = fit$extrapolation_list,
  Sdreport = fit$parameter_estimates$SD,
  Map = fit$tmb_list$Map,
  zlim = NULL,
  ...
)
}
\arguments{
\item{plot_set}{integer-vector defining plots to create
\describe{
  \item{\code{plot_set=1}}{Link-transformed 1st linear predictor, \code{Report$R1_gct}}
  \item{\code{plot_set=2}}{Link-transformed 2nd linear predictor, logged for interpretable plotting scale, \code{log(Report$R2_gct)}}
  \item{\code{plot_set=3}}{Log-predicted density, derived from link-transform of both linear predictors, \code{log(Report$D_gct)}}
  \item{\code{plot_set=6}}{Spatio-temporal variation in 1st linear predictor (e.g., encounter probability when using a conventional delta-model), \code{Report$Epsilon1_gct}}
  \item{\code{plot_set=7}}{Spatio-temporal variation in 2nd linear predictor (e.g., log-positive catch rates  when using a conventional delta-model), \code{Report$Epsilon2_gct}}
  \item{\code{plot_set=8}}{1st linear predictor, \code{Report$P1_gct}}
  \item{\code{plot_set=9}}{2nd linear predictor, \code{Report$P2_gct}}
  \item{\code{plot_set=11}}{Covariates that are included in the model for the 1st linear predictor, \code{Report$X1_gcp}}
  \item{\code{plot_set=12}}{Covariates that are included in the model for the 2nd linear predictor, \code{Report$X2_gcp}}
  \item{\code{plot_set=13}}{Total biomass across all categories (only useful in a multivariate model)}
  \item{\code{plot_set=14}}{Covariate effects on 1st linear predictor, \code{Report$eta1_gct}}
  \item{\code{plot_set=15}}{Covariate effects on 2nd linear predictor, \code{Report$eta2_gct}}
  \item{\code{plot_set=16}}{Spatial variation for 1st linear predictor, \code{Report$Omega1_gc}}
  \item{\code{plot_set=17}}{Spatial variation for 2nd linear predictor, \code{Report$Omega2_gc}}
  \item{\code{plot_set=18}}{Spatially-varying response for density covariates in 1st linear predictor, \code{Report$Xi1_gcp}}
  \item{\code{plot_set=19}}{Spatially-varying response for density covariates in 2nd linear predictor, \code{Report$Xi2_gcp}}
  \item{\code{plot_set=20}}{Spatially-varying response for catchability covariates in 1st linear predictor, \code{Report$Phi1_gk}}
  \item{\code{plot_set=21}}{Spatially-varying response for catchability covariates in 2nd linear predictor, \code{Report$Phi2_gk}}
}}

\item{projargs}{Character passed to \code{\link[sp]{CRS}} and see that documentation for description.
See \url{https://proj.org/operations/projections/index.html} for a list of projections to pass via \code{projargs}.
As a stable default please consider \code{projargs='+proj=natearth +lon_0=0 +units=km'}
where argument \code{+lon_0} allows the user to center eastings on a specified longitude.}

\item{Panel}{Whether to plot years for a given category (\code{Panel="Category"}) or categories for a given year (\code{Panel="Year"})  in each panel figure}

\item{year_labels}{character vector specifying names for labeling times \code{t_i}}

\item{years_to_plot}{integer vector, specifying positions of \code{year_labels} for plotting (used to avoid plotting years with no data, etc.)}

\item{category_names}{character vector specifying names for labeling categories \code{c_i}}

\item{MapSizeRatio}{Default size for each panel}

\item{n_cells}{Integer used to determine the argument \code{cell.size} passed
to \code{\link[raster]{rasterize}} when converting output at extrapolation-grid cells
to a raster prior to plotting mapped outputs using \code{\link{plot_variable}}.  An increased \code{n_cells} results in a decreased
\code{cell.size} and hence higher resolution plots.  The default value for \code{n_cells}
will often need to be modified for a given purpose.}

\item{plot_value}{either \code{plot_value="estimate"} (the default), or a user-specified function that is applied to \code{n_samples} samples from the joint predictive distribution, e.g., to visualize the standard error of a variable by specifying \code{plot_value=sd}}

\item{n_samples}{number of samples from the joint predictive distribution for fixed and random effects.  Default is 100, which is slow.}

\item{country}{optional list of countries to display, e.g. \code{country = c("united states of america", "canada")}. If maps are generating visual artefacts, please try using argument \code{country} to simplify the polygons used to represent land features.}

\item{sample_fixed}{whether to sample fixed and random effects, \code{sample_fixed=TRUE} as by default, or just sample random effects, \code{sample_fixed=FALSE}}

\item{Report}{tagged list of outputs from TMB model via \code{Obj$report()}}

\item{Obj}{Fitted TMB object from package `VAST`, i.e., output from `\code{fit_model(...)$tmb_list$Obj}`}

\item{Sdreport}{Standard deviation outputs from TMB model via \code{sdreport(Obj)}}

\item{zlim}{two numeric values, specifying range for defining bounds of color scale.
If \code{zlim=NULL}, then a constant scale is inferred
from the range of the plotted variable and a color-legend is plotted in the last panel.
If \code{zlim=NA} then a different range is used in each panel from the range of \code{Y_gt[,t]} and a color-legend is plotted in every panel.}

\item{...}{arguments passed to \code{\link{plot_variable}}}
}
\value{
\code{Mat_xt} a matrix (rows: modeled knots; column: modeled year) for plotted output of last element of \code{plot_set}
}
\description{
\code{plot_maps} plots a standard set of diagnostic maps
}
