% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Plot_Overdispersion}
\alias{Plot_Overdispersion}
\title{Copy of plot_overdispersion}
\usage{
Plot_Overdispersion(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{plot_overdispersion}}
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?plot_overdispersion} to see list of arguments and usage
}
