% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantile_residuals.R
\name{plot_quantile_residuals}
\alias{plot_quantile_residuals}
\title{Plot quantile residuals on map}
\usage{
plot_quantile_residuals(
  dharmaRes,
  fit,
  file_name = "quantile_residuals_on_map",
  zlim = NULL,
  n_cells_residuals = NULL,
  year_labels = fit$year_labels,
  years_to_plot = fit$years_to_plot,
  ...
)
}
\arguments{
\item{zlim}{two numeric values, specifying range for defining bounds of color scale.
If \code{zlim=NULL}, then a constant scale is inferred
from the range of the plotted variable and a color-legend is plotted in the last panel.
If \code{zlim=NA} then a different range is used in each panel from the range of \code{Y_gt[,t]} and a color-legend is plotted in every panel.}

\item{n_cells_residuals}{number of raster cells to use when plotting quantile residuals}

\item{year_labels}{character vector specifying names for labeling times \code{t_i}}

\item{years_to_plot}{integer vector, specifying positions of \code{year_labels} for plotting (used to avoid plotting years with no data, etc.)}

\item{...}{arguments passed to \code{\link{plot_variable}}}

\item{output}{from \code{summary.fit_model(x,what="residuals")}}

\item{x}{Output from \code{\link{fit_model}}}
}
\description{
\code{plot_quantile_residuals} shows aggregated quantile residuals at a user-specified spatial resolution

This function plots quantile residuals on a map, and is useful to check for persistent spatial patterns
in those residuals.  For a check for outliers (i.e., the scale of residuals) please see the separate Q-Q plot.
See \code{\link{summary.fit_model}} for more details regarding inputs
}
