% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_range_edge.R
\name{plot_range_edge}
\alias{plot_range_edge}
\title{Plot shifts in range edges}
\usage{
plot_range_edge(
  Sdreport,
  Obj,
  year_labels = NULL,
  years_to_plot = NULL,
  strata_names = NULL,
  category_names = NULL,
  working_dir = getwd(),
  quantiles = c(0.05, 0.95),
  n_samples = 100,
  interval_width = 1,
  width = NULL,
  height = NULL,
  calculate_relative_to_average = FALSE,
  seed = 123456,
  ...
)
}
\arguments{
\item{Sdreport}{Standard deviation outputs from TMB model via \code{sdreport(Obj)}}

\item{Obj}{Fitted TMB object from package `VAST`, i.e., output from `\code{fit_model(...)$tmb_list$Obj}`}

\item{year_labels}{character vector specifying names for labeling times \code{t_i}}

\item{years_to_plot}{integer vector, specifying positions of \code{year_labels} for plotting (used to avoid plotting years with no data, etc.)}

\item{strata_names}{names for spatial strata}

\item{category_names}{names for categories (if using package \code{`VAST`})}

\item{working_dir}{Directory for plots}

\item{quantiles}{vector specifying quantiles to use for calculating range edges}

\item{n_samples}{number of samples from the joint predictive distribution for fixed and random effects.  Default is 100, which is slow.}

\item{interval_width}{width for confidence intervals}

\item{width}{plot width in inches}

\item{height}{plot height in inches}

\item{calculate_relative_to_average}{Boolean, whether to calculate edge in UTM coordinates (default),
or instead calculate relative to median across all years. The latter reduces standard errors,
and is appropriate when checking significance for comparison across years for a single species.
The former (default) is appropriate for checking significance for comparison across species.}

\item{seed}{integer used to set random-number seed when sampling variables, as passed to \code{set.seed(.)}}

\item{...}{Other inputs to `par()`}
}
\description{
\code{plot_range_edge} plots range edges
}
\references{
For details regarding multivariate index standardization and expansion see \url{https://doi.org/10.22541/au.160331933.33155622/v1}
}
