% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_range_index.R
\name{plot_range_index}
\alias{plot_range_index}
\title{Plot shifts in distribution and area occupied}
\usage{
plot_range_index(
  Sdreport,
  Report,
  TmbData,
  year_labels = NULL,
  years_to_plot = NULL,
  strata_names = NULL,
  PlotDir = getwd(),
  FileName_COG = "center_of_gravity.png",
  FileName_EffArea = "Effective_Area.png",
  Znames = rep("", ncol(TmbData$Z_gm)),
  use_biascorr = TRUE,
  category_names = NULL,
  interval_width = 1,
  ...
)
}
\arguments{
\item{Sdreport}{TMB output from `TMB::sdreport(Obj)`}

\item{Report}{Reporting output from `Obj$report()`}

\item{TmbData}{Formatted data inputs, from `VAST::Data_Fn(...)`}

\item{year_labels}{character vector specifying names for labeling times \code{t_i}}

\item{years_to_plot}{integer vector, specifying positions of \code{year_labels} for plotting (used to avoid plotting years with no data, etc.)}

\item{strata_names}{names for spatial strata}

\item{PlotDir}{Directory for plots}

\item{FileName_COG}{Full filename (including directory) for center-of-gravity plot}

\item{FileName_EffArea}{Full filename (including directory) for center-of-gravity plot}

\item{Znames}{Names for center-of-gravity summary statistics}

\item{use_biascorr}{Boolean, whether to use bias-corrected estimates if available}

\item{category_names}{names for categories (if using package \code{`VAST`})}

\item{interval_width}{width for confidence intervals}

\item{...}{Other inputs to `par()`}

\item{FileName_Area}{Full filename (including directory) for center-of-gravity plot}
}
\value{
Return Tagged list of output
\describe{
  \item{COG_Table}{table of center-of-gravity ("COG") estimates by year}
  \item{KernelArea_Table}{table of kernel-area approximation to area occupied, estimated by year}
  \item{EffectiveArea_Table}{table of effective-area approximation to area occupied, estimated by year, recommended over \code{KernelArea_Table}}
}
}
\description{
\code{plot_range_index} plots center-of-gravity, kernel-area occupied, and effective-area occupied
}
\references{
For details regarding center of gravity see \url{https://doi.org/10.1111/2041-210X.12567}

For details regarding effective area occupied see \url{https://doi.org/10.1098/rspb.2016.1853}
}
