% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residual_semivariance.R
\name{plot_residual_semivariance}
\alias{plot_residual_semivariance}
\title{Plot semivariance for quantile residuals}
\usage{
plot_residual_semivariance(
  fit,
  dharma_raster,
  dharmaRes,
  file_name = "quantile_residuals_semivariance",
  working_dir = getwd()
)
}
\description{
\code{plot_residual_semivariance} calculates the spatial and temporal semivariance from standard-normal quantile residuals

This function plots takes quantile residuals, converts them to a standard-normal distribution,
     fits the two-dimensional semi-variance in space and time, and plots this.
}
