% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot Pearson residuals on map}
\usage{
plot_residuals(
  Lat_i,
  Lon_i,
  TmbData,
  Report,
  Q,
  projargs = "+proj=longlat",
  working_dir = paste0(getwd(), "/"),
  spatial_list,
  extrapolation_list,
  Year_Set = NULL,
  Years2Include = NULL,
  zrange,
  ...
)
}
\arguments{
\item{Lat_i}{Latitude for every observation \code{i}}

\item{Lon_i}{Longitude for every observation \code{i}}

\item{TmbData}{a tagged list of data inputs generated by \code{make_data}}

\item{Report}{tagged list of outputs from TMB model via \code{Obj$report()}}

\item{Q}{Output from \code{QQ_Fn}}

\item{projargs}{Character passed to \code{\link[sp]{CRS}} and see that documentation for description.
See \url{https://proj.org/operations/projections/index.html} for a list of projections to pass via \code{projargs}.
As a stable default please consider \code{projargs='+proj=natearth +lon_0=0 +units=km'}
where argument \code{+lon_0} allows the user to center eastings on a specified longitude.}

\item{zrange}{lower and upper bounds for plotted value of Pearson residuals, used to change plotting scale to eliminate influence of outliers on plotting scale}

\item{...}{arguments passed to \code{FisshStatsUtils::plot_variable}}

\item{savdir}{directory to use when saving results}
}
\value{
A tagged list of Pearson residuals
\describe{
  \item{Q1_xt}{Matrix of average residuals for encounter/non-encounter component by site \code{x} and year \code{t}}
  \item{Q2_xt}{Matrix of average residuals for positive-catch-rate component by site \code{x} and year \code{t}}
}
}
\description{
\code{plot_residuals} shows average Pearson residual for every knot for encounter probability and positive catch rate components
}
