% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{Plot results}
\usage{
plot_results(
  fit,
  settings = fit$settings,
  plot_set = 3,
  working_dir = getwd(),
  year_labels = fit$year_labels,
  years_to_plot = fit$years_to_plot,
  category_names = fit$category_names,
  strata_names = fit$strata_names,
  use_biascorr = TRUE,
  map_list = NULL,
  check_residuals = TRUE,
  cluster_results = TRUE,
  projargs = "+proj=longlat",
  zrange,
  n_samples = 100,
  calculate_relative_to_average = FALSE,
  type = 1,
  n_cells = NULL,
  n_cells_residuals = NULL,
  RotationMethod = "PCA",
  quantiles = c(0.05, 0.5, 0.95),
  similarity_metric = c("hclust", "Correlation", "Dissimilarity", "Covariance")[1],
  ...
)
}
\arguments{
\item{fit}{Output from \code{fit_model}}

\item{settings}{Output from \code{\link{make_settings}}}

\item{plot_set}{integer-vector defining plots to create
\describe{
  \item{\code{plot_set=1}}{Link-transformed 1st linear predictor, \code{Report$R1_gct}}
  \item{\code{plot_set=2}}{Link-transformed 2nd linear predictor, logged for interpretable plotting scale, \code{log(Report$R2_gct)}}
  \item{\code{plot_set=3}}{Log-predicted density, derived from link-transform of both linear predictors, \code{log(Report$D_gct)}}
  \item{\code{plot_set=6}}{Spatio-temporal variation in 1st linear predictor (e.g., encounter probability when using a conventional delta-model), \code{Report$Epsilon1_gct}}
  \item{\code{plot_set=7}}{Spatio-temporal variation in 2nd linear predictor (e.g., log-positive catch rates  when using a conventional delta-model), \code{Report$Epsilon2_gct}}
  \item{\code{plot_set=8}}{1st linear predictor, \code{Report$P1_gct}}
  \item{\code{plot_set=9}}{2nd linear predictor, \code{Report$P2_gct}}
  \item{\code{plot_set=11}}{Covariates that are included in the model for the 1st linear predictor, \code{Report$X1_gcp}}
  \item{\code{plot_set=12}}{Covariates that are included in the model for the 2nd linear predictor, \code{Report$X2_gcp}}
  \item{\code{plot_set=13}}{Total biomass across all categories (only useful in a multivariate model)}
  \item{\code{plot_set=14}}{Covariate effects on 1st linear predictor, \code{Report$eta1_gct}}
  \item{\code{plot_set=15}}{Covariate effects on 2nd linear predictor, \code{Report$eta2_gct}}
  \item{\code{plot_set=16}}{Spatial variation for 1st linear predictor, \code{Report$Omega1_gc}}
  \item{\code{plot_set=17}}{Spatial variation for 2nd linear predictor, \code{Report$Omega2_gc}}
  \item{\code{plot_set=18}}{Spatially-varying response for density covariates in 1st linear predictor, \code{Report$Xi1_gcp}}
  \item{\code{plot_set=19}}{Spatially-varying response for density covariates in 2nd linear predictor, \code{Report$Xi2_gcp}}
  \item{\code{plot_set=20}}{Spatially-varying response for catchability covariates in 1st linear predictor, \code{Report$Phi1_gk}}
  \item{\code{plot_set=21}}{Spatially-varying response for catchability covariates in 2nd linear predictor, \code{Report$Phi2_gk}}
}}

\item{working_dir}{Directory for plots}

\item{year_labels}{character vector specifying names for labeling times \code{t_i}}

\item{years_to_plot}{integer vector, specifying positions of \code{year_labels} for plotting (used to avoid plotting years with no data, etc.)}

\item{category_names}{character vector specifying names for labeling categories \code{c_i}}

\item{strata_names}{names for spatial strata}

\item{map_list}{output from \code{\link{make_map_info}}, which is included in output of \code{\link{plot_results}}}

\item{check_residuals}{Boolean indicating whether to run or skip residual diagnostic plots (which can be slow as currently implemented)}

\item{cluster_results}{Boolean whether to run \code{\link{plot_clusters}}, which can be slow for analyses involving a large number
of extrapolation-grid cells}

\item{projargs}{Character passed to \code{\link[sp]{CRS}} and see that documentation for description.
See \url{https://proj.org/operations/projections/index.html} for a list of projections to pass via \code{projargs}.
As a stable default please consider \code{projargs='+proj=natearth +lon_0=0 +units=km'}
where argument \code{+lon_0} allows the user to center eastings on a specified longitude.}

\item{n_samples}{number of samples from the joint predictive distribution for fixed and random effects.  Default is 100, which is slow.}

\item{calculate_relative_to_average}{Boolean, whether to calculate edge in UTM coordinates (default),
or instead calculate relative to median across all years. The latter reduces standard errors,
and is appropriate when checking significance for comparison across years for a single species.
The former (default) is appropriate for checking significance for comparison across species.}

\item{type}{integer stating what type of simulation to use from the following options:
\itemize{
\item \code{type=1} is a "measurement error" or "conditional" simulator that simulates new data conditional upon estimated fixed and random effects.
\item \code{type=2} is an "unconditional" simulator that simulates new random effects conditional upon fixed effects
(but not otherwise conditioning upon original data), and new data conditional upon both.
\item \code{type=3} simulates new fixed and random effects from the joint precision matrix (i.e., conditioning upon the original data), and new data conditional upon these values.
\item \code{type=4} simulates new random effects from the internal Hessian matrix evaluated at the MLE (i.e., conditional on fixed effects estimates and the original data),
and new data conditional upon these values.
}}

\item{n_cells}{Integer used to determine the argument \code{cell.size} passed
to \code{\link[raster]{rasterize}} when converting output at extrapolation-grid cells
to a raster prior to plotting mapped outputs using \code{\link{plot_variable}}.  An increased \code{n_cells} results in a decreased
\code{cell.size} and hence higher resolution plots.  The default value for \code{n_cells}
will often need to be modified for a given purpose.}

\item{n_cells_residuals}{number of raster cells to use when plotting quantile residuals}

\item{RotationMethod}{Method used for rotation when visualing factor decomposition results,
Options: "PCA" (recommended) or "Varimax"}

\item{quantiles}{vector specifying quantiles to use for calculating range edges}

\item{similarity_metric}{approach used to visualize similarity among years/categories
resulting from estimated loadings matrices.  Available options include
\code{"hclust", "Correlation", "Dissimilarity", "Covariance"}}

\item{...}{additional settings to pass to \code{\link{plot_maps}}}
}
\value{
Invisibly returns a tagged list of outputs generated by standard plots. See linked functions for details
\describe{
  \item{\code{dharmaRes}}{Output from \code{\link{summary.fit_model}}, representing quantile residuals calculated using package DHARMa}
  \item{\code{Enc_prob}}{Output from \code{\link{plot_encounter_diagnostic}}}
  \item{\code{Index}}{Output from \code{\link{plot_biomass_index}}}
  \item{\code{Proportions}}{Output from \code{\link{calculate_proportion}}}
  \item{\code{Range}}{Output from \code{\link{plot_range_index}}}
  \item{\code{Dens_xt}}{Output from \code{\link{plot_maps}}}
  \item{\code{Edge}}{Output from \code{\link{plot_range_edge}}}
  \item{\code{Factors}}{Output from \code{\link{plot_factors}}}
  \item{\code{Clusters}}{Output from \code{\link{plot_clusters}}}
}
}
\description{
\code{plot_results} plots diagnostics, results, and indices for a given fitted model
}
\details{
This function takes a fitted VAST model and generates a standard set of diagnostic and visualization plots.
It does this by calling a series of mid-level plotting functions; see list of functions in Value section of documentation.

In particular, for making customized maps of output please see \code{\link{plot_variable}}
}
\seealso{
\code{\link[VAST]{VAST}} for general documentation, \code{\link{make_settings}} for generic settings, \code{\link{fit_model}} for model fitting, and \code{\link{plot_results}} for generic plots

Other wrapper functions: 
\code{\link{fit_model}()},
\code{\link{make_settings}()}
}
\concept{wrapper functions}
