% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_similarity.R
\name{plot_similarity}
\alias{plot_similarity}
\title{Plot similarity resulting from estimated covariance}
\usage{
plot_similarity(
  fit,
  year_labels = fit$year_labels,
  category_names = fit$category_names,
  similarity_metric = c("hclust", "Correlation", "Dissimilarity", "Covariance")[1],
  working_dir = getwd(),
  file_name = similarity_metric,
  panel_size = 3,
  Res = 200,
  Format = "png"
)
}
\arguments{
\item{fit}{output from \code{fit_model}}

\item{similarity_metric}{approach used to visualize similarity among years/categories
resulting from estimated loadings matrices.  Available options include
\code{"hclust", "Correlation", "Dissimilarity", "Covariance"}}
}
\description{
\code{plot_similarity} plots metrics of similarity derived from estimated covariance matrices
}
