% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot timeseries estimates}
\usage{
plot_timeseries(
  x,
  y,
  y_sd,
  ybounds,
  fn = lines,
  ylim = NULL,
  bounds_type = "whiskers",
  bounds_args = list(),
  interval_width = 1,
  connect_ascending = TRUE,
  ...
)
}
\arguments{
\item{x}{numeric-vector of x-coordinates to plot}

\item{y}{numeric-vector of y-coordinates to plot}

\item{ybounds}{alternate specification of y-coordinate whiskers; useful to avoid normal approximation}

\item{fn}{what plotting function to use; default \code{fn=lines} does not create a new plotting window}

\item{bounds_type}{which type either \code{"whiskers"} or \code{"shading"}}

\item{bounds_args}{tagged list of additional arguments to pass to \code{\link[graphics]{lines}} or \code{\link[graphics]{polygon}}}

\item{interval_width}{width of interval in normal approximation; only used when specifying \code{ysd} without \code{ybounds}}

\item{ysd}{numeric-vector of standard errors in y-coordinates, used in normal approximation to interval}
}
\description{
\code{plot_timeseries} plots a timeseries with standard errors shown as whiskers or a shaded polygon
}
