% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable.R
\name{plot_variable}
\alias{plot_variable}
\title{Plot results on a multi-panel figure of maps}
\usage{
plot_variable(
  Y_gt,
  map_list = NULL,
  panel_labels = NULL,
  projargs = "+proj=longlat",
  map_resolution = "medium",
  file_name = "density",
  working_dir = getwd(),
  Format = "png",
  Res = 200,
  add = FALSE,
  outermargintext = c("Eastings", "Northings"),
  zlim = NULL,
  col = viridisLite::viridis,
  mar = c(0, 0, 2, 0),
  oma = c(4, 4, 0, 0),
  legend_x = c(0, 0.05),
  legend_y = c(0.05, 0.45),
  cex.legend = 1,
  mfrow = NULL,
  land_color = "grey",
  n_cells = NULL,
  xlim = NULL,
  ylim = NULL,
  country = NULL,
  contour_nlevels = 0,
  fun = mean,
  format = "sf",
  cex.points = 1,
  legend_digits = 1,
  ...
)
}
\arguments{
\item{Y_gt}{matrix where values for every column are plotted as a map in a panel-figure with one panel per column}

\item{map_list}{output from \code{\link{make_map_info}}, which is included in output of \code{\link{plot_results}}}

\item{panel_labels}{vector defining titles to use for each panel; defaults to blank}

\item{projargs}{Character passed to \code{\link[sp]{CRS}} and see that documentation for description.
See \url{https://proj.org/operations/projections/index.html} for a list of projections to pass via \code{projargs}.
As a stable default please consider \code{projargs='+proj=natearth +lon_0=0 +units=km'}
where argument \code{+lon_0} allows the user to center eastings on a specified longitude.}

\item{add}{boolean indicating whether to add plot to an existing panel figure, or to define a new panel figure}

\item{outermargintext}{vector defining text to plot in outer margins of panel figure}

\item{zlim}{two numeric values, specifying range for defining bounds of color scale.
If \code{zlim=NULL}, then a constant scale is inferred
from the range of the plotted variable and a color-legend is plotted in the last panel.
If \code{zlim=NA} then a different range is used in each panel from the range of \code{Y_gt[,t]} and a color-legend is plotted in every panel.}

\item{legend_x}{two numeric values (generally between 0 and 1, but slightly lower/higher values
generate colorbars that are just outside the plotting window) giving left and right-hand location of color legend}

\item{legend_y}{two numeric values (see description for legend_x) giving bottom and top location of color legend}

\item{land_color}{color for filling in land (use \code{land_color=rgb(0,0,0,alpha=0)} for transparent land)}

\item{n_cells}{Integer used to determine the argument \code{cell.size} passed
to \code{\link[raster]{rasterize}} when converting output at extrapolation-grid cells
to a raster prior to plotting mapped outputs using \code{\link{plot_variable}}.  An increased \code{n_cells} results in a decreased
\code{cell.size} and hence higher resolution plots.  The default value for \code{n_cells}
will often need to be modified for a given purpose.}

\item{country}{a character vector of country names.}

\item{contour_nlevels}{number of levels used when adding contour lines, passed to \code{\link[graphics]{contour}} as argument \code{nlevels}}

\item{fun}{function or character. To determine what values to assign to cells that are covered by multiple spatial features. You can use functions such as \code{min, max}, or \code{mean}, or one of the following character values: \code{'first'}, \code{'last'}, \code{'count'}. The default value is \code{'last'}. In the case of SpatialLines*, \code{'length'} is also allowed (currently for planar coordinate systems only).


If \code{x} represents points, \code{fun} must accept a \code{na.rm} argument, either explicitly or through the ellipses ('dots'). This means that \code{fun=length} fails, but \code{fun=function(x,...)length(x)} works, although it ignores the \code{na.rm} argument. To use the \code{na.rm} argument you can use a function like this: \code{fun=function(x, na.rm){if (na.rm) length(na.omit(x)) else (length(x)}}, or use a function that removes \code{NA} values in all cases, like this function to compute the number of unique values per grid cell "richness": \code{fun=function(x, ...) {length(unique(na.omit(x)))} }. If you want to count the number of points in each grid cell, you can use \code{ fun='count'} or  \code{fun=function(x,...){length(x)}}.

You can also pass multiple functions using a statement like \code{fun=function(x, ...) c(length(x),mean(x))}, in which case the returned object is a RasterBrick (multiple layers).
}

\item{format}{Character specifying whether to plot maps as a raster (the default), \code{format="raster"}, or as points \code{format="points"}}

\item{cex.points}{Numeric specifying the size of bullets when \code{format="points"}}

\item{legend_digits}{number of digits (i.e., value passed to \code{round}) when creating labels for colorbar legend}

\item{...}{arguments passed to \code{par}}
}
\description{
\code{plot_variable} plots a map and fills in regions with colors to represent intensity in an areal-interpretion of model results

If maps are generating visual artefacts, please try using argument \code{country} to simplify the polygons used to represent land features.

In some sampling designs and/or analyses, it is not appropriate to extrapolate beyond samping stations.  In these cases,
the analyst may define the extrapolation-grid to only represent sampling stations, and then use \code{format="points"}
to restrict mapping to bullets at those locations.
}
