% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollock_cohorts}
\alias{pollock_cohorts}
\title{Data to demonstrate and test cohort effects in age-structured SDM}
\usage{
data(pollock_cohorts)
}
\description{
Data sufficient to demonstrate how spatially varying coefficients (SVCs)
     are used to represent cohort effects in a joint (age-structured)
     species distribution model.  Specifically using processed data
     from the Alaska Fisheries Science Center bottom trawl survey
     that includes a density-dependent correction (based on subsampled acoustics)
     and allocated to ages using an age-length-key (ALK), where both were
     conducted by Stan Kotwicki in support
}
\details{
\itemize{
  \item sampling_data data-frame of tow-specific biomass
}
}
\references{
Stevenson, D. E., Kotwicki, S., Thorson, J. T., Correa, G. M., and Buckley, T. 2022.
The influence of age and cohort on the distribution of walleye pollock (Gadus chalcogrammus)
in the eastern Bering Sea. Canadian Journal of Fisheries and Aquatic Sciences.
<https://doi.org/10.1139/cjfas-2021-030>
}
\author{
Stan Kotwicki
}
\keyword{data}
