% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{print.make_data}
\alias{print.make_data}
\title{Print data}
\usage{
\method{print}{make_data}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{make_data}}}

\item{...}{Not used}
}
\description{
Print data fitted by \code{\link{VAST}}
}
