% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_extrapolation_info.R
\name{print.make_extrapolation_info}
\alias{print.make_extrapolation_info}
\title{Extract extrapolation-grid}
\usage{
\method{print}{make_extrapolation_info}(x, quiet = FALSE, ...)
}
\arguments{
\item{x}{Output from \code{\link{make_extrapolation_info}}}

\item{quiet}{Boolean indicating whether to print to terminal or not}

\item{...}{Not used}
}
\value{
Data frame of Latitude and Longitude for each extrapolation-grid cell
}
\description{
Extract extrapolation-grid used for spatial inference
}
