% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model.R
\name{print.make_model}
\alias{print.make_model}
\title{Print model object}
\usage{
\method{print}{make_model}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{make_model}}}

\item{...}{Not used}
}
\description{
Print model object generated by \code{\link{VAST}}
}
