% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spatial_info.R
\name{print.make_spatial_info}
\alias{print.make_spatial_info}
\title{Print spatial information}
\usage{
\method{print}{make_spatial_info}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{make_spatial_info}}}

\item{...}{Not used}
}
\description{
Print spatial representation used for model
}
