% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_coordinates.R
\name{project_coordinates}
\alias{project_coordinates}
\title{Convert from Lat-Long to UTM}
\usage{
project_coordinates(
  X,
  Y,
  projargs = NA,
  origargs = "+proj=longlat +datum=WGS84",
  zone = NA,
  flip_around_dateline = FALSE,
  silent = FALSE
)
}
\arguments{
\item{X}{vector of horizontal-coordinates (e.g., longitude by default)}

\item{Y}{vector of vertical-coordinates (e.g., latitude by default)}

\item{projargs}{A character string of projection arguments; the arguments must be entered exactly as in the PROJ.4 documentation; if the projection is unknown, use \code{as.character(NA)}, it may be missing or an empty string of zero length and will then set to the missing value. With \pkg{rgdal} built with PROJ >= 6 and GDAL >= 3, the \code{+init=} key may only be used with value \code{epsg:<code>}. From \pkg{sp} version 1.4-4, the string associated with the SRS_string argument may be entered as-is and will be set as SRS_string if the projargs argument does not begin with a \code{+} (suggested by Mikko Vihtakari).}

\item{zone}{DEPRECATED INPUT; UTM zone (integer between 1 and 60) or alphanumeric CRS code used by package rgdal to convert latitude-longitude coordinates to projection in kilometers; \code{zone=NA} uses UTM and automatically detects the appropriate zone}

\item{flip_around_dateline}{DEPRECATED INPUT; boolean specifying whether to flip Lat-Lon locations around the dateline, and then retransform back (only useful if Lat-Lon straddle the dateline)}
}
\value{
A data frame with the following columns
\describe{
  \item{X}{The horizonal coordinates after projection (e.g., Eastings in UTM kilometers by default)}
  \item{Y}{The vertical coordinates after projection (e.g., Northings in UTM kilometers by default)}
}
}
\description{
\code{project_coordinates} converts from Latitude-Longitude to Universal Transverse Mercator projections for a given location
}
