% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rerun_Fn.R
\name{Rerun_Fn}
\alias{Rerun_Fn}
\title{Explore counter-factual scenario}
\usage{
Rerun_Fn(parhat0, turnoff_pars, loc_x, calculate_COG = TRUE, figname = NULL,
  Map = "generate", MapDetails_List = NULL,
  year_set = 1:ncol(parhat_mod[["beta1_ct"]]),
  c_set = 1:nrow(parhat_mod[["beta1_ct"]]), ...)
}
\arguments{
\item{parhat0}{parameter values to use by default (presumably maximum-likelihood estimates)}

\item{turnoff_pars}{character-vector of parameters to turn off}

\item{loc_x}{location for each knot used to calculate center-of-gravity for counter-factual runs with colnames c('E_km','N_km','Lon','Lat')}

\item{calculate_COG}{Boolean whether to calculate COG for each run}

\item{figname}{name for figure to plot density in counter-factual scenario}

\item{Map}{OPTIONAL, a tagged list of parameters to either mirror or turn off}

\item{MapDetails_List}{output from \code{SpatialDeltaGLMM::MapDetails_Fn}}

\item{year_set}{set of parameters to include}

\item{c_set}{set of categories to include}

\item{...}{additional arguments passed to \code{VAST::Build_TMB_Fn}}
}
\value{
Tagged list
\describe{
  \item{Report}{Report output for counter-factual run}
  \item{NewBuild_List}{Output from \code{VAST::Build_TMB_Fn} using counter-factual parameters}
}
}
\description{
\code{Rerun_Fn} re-builds a model while fixing a subset of parameters at zero
}

