% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_factors.R
\name{rotate_factors}
\alias{rotate_factors}
\title{Rotate results}
\usage{
rotate_factors(
  L_pj = NULL,
  Psi_sjt = NULL,
  Cov_pp = NULL,
  Psi_spt = NULL,
  RotationMethod = "PCA",
  testcutoff = 1e-10,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{L_pj}{Loadings matrix for `p` categories and `j` factors (calculated from \code{Cov_pp} if it is provided)}

\item{Psi_sjt}{Array of factors (1st dimension: spatial knots;  2nd dimension: factors;  3rd dimension:  time)}

\item{Cov_pp}{Covariance calculated from loadings matrix.
if \code{Cov_pp} is provided and \code{L_pj=NULL}, then \code{L_pj} is calculated from \code{Cov_pp}}

\item{Psi_spt}{Array of projected factors (1st dimension: spatial knots;  2nd dimension: categories;  3rd dimension:  time)
if \code{Psi_spt} is provided and \code{Psi_sjt=NULL}, then \code{Psi_sjt} is calculated from
\code{Psi_spt} and \code{L_pj}}

\item{RotationMethod}{Method used for rotation when visualing factor decomposition results,
Options: "PCA" (recommended) or "Varimax"}

\item{testcutoff}{tolerance for numerical rounding when confirming that rotation doesn't effect results}
}
\value{
tagged list of outputs
\describe{
  \item{L_pj_rot}{Loadings matrix after rotation}
  \item{Psi_rot}{Factors after rotation}
  \item{Hinv}{Object used for rotation}
  \item{L_pj}{Loadings matrix}
}
}
\description{
\code{rotate_factors} rotates results from a factor model
}
