% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate new dynamics and sampling data}
\usage{
simulate_data(fit, type = 1, random_seed = NULL)
}
\arguments{
\item{fit}{output from \code{\link{fit_model}}}

\item{type}{integer stating what type of simulation to use from the following options:
\itemize{
\item \code{type=1} is a "measurement error" or "conditional" simulator that simulates new data conditional upon estimated fixed and random effects.
\item \code{type=2} is an "unconditional" simulator that simulates new random effects conditional upon fixed effects
(but not otherwise conditioning upon original data), and new data conditional upon both.
\item \code{type=3} simulates new fixed and random effects from the joint precision matrix (i.e., conditioning upon the original data), and new data conditional upon these values.
\item \code{type=4} simulates new random effects from the internal Hessian matrix evaluated at the MLE (i.e., conditional on fixed effects estimates and the original data),
and new data conditional upon these values.
}}

\item{random_seed}{integer passed to \code{set.seed}, where the default value \code{random_seed=NULL} resets the random-number seed.}
}
\value{
Report object containing new data and population variables including
\describe{
  \item{b_i}{New simulated data}
  \item{D_gcy}{Density for each grid cell g, category c, and year y}
  \item{Index_cyl}{Index of abundance for each category c, year y, and stratum l}
}
}
\description{
\code{simulate_data} conducts a parametric bootstrap to simulate new data and potentially simulate new population dynamics and associated variables

Simulate new data given various potential procedures to propagate uncertainty about parameters.

Using \code{sample_fixed=TRUE} (the default) in \code{\link{sample_variable}} is similar to using \code{type=3} in \code{\link{simulate_data}}, while
      using \code{sample_fixed=TRUE} in \code{\link{sample_variable}} is similar to using \code{type=4} in \code{\link{simulate_data}}.
      Sampling fixed effects will sometimes cause numerical under- or overflow (i.e., output values of \code{NA}) in cases when
      variance parameters are estimated imprecisely.  In these cases, the multivariate normal approximation being used is a poor
      representation of the tail probabilities, and results in some samples with implausibly high (or negative) variances,
      such that the associated random effects then have implausibly high magnitude.
}
