% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stream_network_eel_example}
\alias{stream_network_eel_example}
\title{Data to demonstrate and test stream-network options}
\usage{
data(stream_network_eel_example)
}
\description{
Data sufficient to demonstrate stream-network options in VAST, using \code{Method="Stream_network"},
 and also helpful to do integrated testing of this feature
}
\details{
\itemize{
  \item observations data-frame of biological sampling data
  \item network data-frame defining stream-network connectivity, used to define Ornstein-Uhlenbeck covariance function
  \item habitat data-frame containing density covariates
}
}
\author{
Merrill Rudd
}
\keyword{data}
