% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_units.R
\name{strip_units}
\alias{strip_units}
\title{Remove units from object}
\usage{
strip_units(x)
}
\description{
\code{strip_units} facilitates logical comparisons and other generic operators by stripping units from an object but otherwise leaving it unchanged
}
