% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{summary.fit_model}
\alias{summary.fit_model}
\title{Extract summary of spatial estimates}
\usage{
\method{summary}{fit_model}(
  x,
  what = "density",
  n_samples = 250,
  working_dir = NULL,
  type = 1,
  random_seed = NULL,
  form = NULL,
  category_names = x$category_names,
  year_labels = x$year_labels,
  ...
)
}
\arguments{
\item{x}{Output from \code{\link{fit_model}}}

\item{what}{String indicating what to summarize; options are `density`, `index` or `residuals`}

\item{n_samples}{Number of samples used when \code{what="residuals"}}

\item{type}{integer stating what type of simulation to use from the following options:
\itemize{
\item \code{type=1} is a "measurement error" or "conditional" simulator that simulates new data conditional upon estimated fixed and random effects.
\item \code{type=2} is an "unconditional" simulator that simulates new random effects conditional upon fixed effects
(but not otherwise conditioning upon original data), and new data conditional upon both.
\item \code{type=3} simulates new fixed and random effects from the joint precision matrix (i.e., conditioning upon the original data), and new data conditional upon these values.
\item \code{type=4} simulates new random effects from the internal Hessian matrix evaluated at the MLE (i.e., conditional on fixed effects estimates and the original data),
and new data conditional upon these values.
}}

\item{random_seed}{integer passed to \code{set.seed}, where the default value \code{random_seed=NULL} resets the random-number seed.}

\item{form}{Optional predictor against which the residuals should be plotted. Default is to used the predicted(simulationOutput).}

\item{...}{additional arguments passed to \code{\link[DHARMa]{plotResiduals}} when \code{what="residuals"}}
}
\description{
\code{summary.fit_model} extracts commonly used quantities derived from a fitted VAST model
}
\details{
\code{summary.fit_model} faciliates common queries for model output including:
\itemize{
\item \code{what="density"} returns a tagged list containing element \code{Density_dataframe},
which lists the estimated density for every Latitude-Longitude-Year-Category combination
for every modelled location in the extrapolation-grid.
\item \code{what="residuals"} returns a DHARMa object containing PIT residuals;
See details section for more information.
}

For calculating residuals, the function calls package \code{\link[DHARMa]{DHARMa}}
to create a diagnostic object for simulation-based quantile residuals.
It specifically simulates replicated data sets from the predictive distribution of data
conditional on estimated fixed and random effects. It then
calculates probability-integral-transform (PIT) residuals from the observed and simulated values.
It then replaces the automatically calculated residuals in the DHARMa object with these these PIT residuals,
so that DHARMa can be used to plot those PIT residuals. PIT residuals are used because the original DHARMa calculations
are not correct when using a delta-model (due to additional jittered values added by DHARMa when detecting multiple 0-valued observations), hence
the need to call this function to correctly calculate PIT residuals for a delta-model.

Note that \code{summary(fit, ..., type=0} uses \code{\link{oneStepPredict_deltaModel}} to calculate one-step-ahead
residuals.  These are probably the most appropriate method for evaluating residuals, but are also *very* slow to calculate relative
to other methods.
}
\seealso{
\code{\link{plot_quantile_residuals}} to plot output of \code{summary.fit_model(x,what="residuals")}
}
