% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Summary.R
\name{Summary}
\alias{Summary}
\title{Summarize and explain results}
\usage{
Summary(obj)
}
\arguments{
\item{obj}{the TMB object after it has been previously optimized}
}
\value{
Tagged list of useful output
\describe{
  \item{parhat}{a list of fixed-effect estimates, with same format as starting values}
  \item{D_xct}{an array (dimensions: knots; categories, e.g., species/age;  year) of predicted density (doesn't include bias-correction)}
}
}
\description{
\code{Summary} is intended to help document results, by providing explanation for outputs from \code{obj$report()}
}
\details{
Please post on the GitHub issue tracker if you'd like for another specific output to be added and documented here
}

