% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBAIC.R
\name{TMBAIC}
\alias{TMBAIC}
\title{Calculate marginal AIC for a fitted model}
\usage{
TMBAIC(opt, p = 2, n = Inf)
}
\arguments{
\item{opt}{the output from \code{nlminb} or \code{optim}}

\item{p}{the penalty on additional fixed effects (default=2, for AIC)}

\item{n}{the sample size, for use in AICc calculation (default=Inf, for which AICc=AIC)}
}
\value{
AIC, where a parsimonious model has a AIC relative to other candidate models
}
\description{
\code{TMBAIC} calculates AIC for a given model fit
}
