% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VAST.R
\docType{package}
\name{VAST}
\alias{VAST}
\alias{VAST-package}
\title{Vector autoregressive spatio-temporal model}
\description{
VAST is multivariate generalization of package `SpatialDeltaGLMM`, and can analyze data for multiple species, ages, size-classes, etc.
}
\details{
Please see examples folder on github \url{http://github.com/james-thorson/VAST/tree/master/examples} for up-to-date examples
}

