% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_epsilon.R
\name{apply_epsilon}
\alias{apply_epsilon}
\title{Custom epsilon-correct method}
\usage{
apply_epsilon(
  fit,
  ADREPORT_name = "Index_ctl",
  eps_name = "eps_Index_ctl",
  inner.control = list(sparse = TRUE, lowrank = TRUE, trace = FALSE)
)
}
\arguments{
\item{fit}{output from \code{\link{fit_model}}, specifically using
slotes \code{tmb_list}, \code{input_args}, \code{parameter_estimates$SD}}

\item{ADREPORT_name}{string indicating name of ADREPORT'ed variable}

\item{eps_name}{string indicating name of PARAMETER used internally by TMB
for calculating desired gradient}

\item{inner.control}{List controlling inner optimization.}
}
\value{
Standard output from \code{\link[TMB]{sdreport}}, but with slot
        \code{x$unbiased} added if needed, and adding or replacing values for
        \code{x$unbiased$value} corresponding to \code{ADREPORT_name}
}
\description{
\code{apply_epsilon} uses updates to TMB to implement a faster calculation for epsilon-correction
}
