% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{bilinear_interp}
\alias{bilinear_interp}
\title{Calculate weights for bilinear interpolation}
\usage{
bilinear_interp(xyz1, xyz2, xyz3, xypred)
}
\arguments{
\item{xyz1, }{2D location and height of first neighbor}

\item{xyz2, }{2D location and height of second neighbor}

\item{xyz3, }{2D location and height of third neighbor}

\item{xypred, }{2D location of location to interpolate}
}
\value{
Tagged list of useful output
\describe{
  \item{zpred}{height at location \code{xypred}}
  \item{phi}{Coefficients for interpolation}
}
}
\description{
\code{bilinear_interp} calculates three coefficients, used to calculate height via blinear interpolation from height and location of three neighbors
}
