% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_TMB_Fn.R
\name{Build_TMB_Fn}
\alias{Build_TMB_Fn}
\title{Build TMB object for VAST model}
\usage{
Build_TMB_Fn(TmbData, Version, RhoConfig = c(Beta1 = 0, Beta2 = 0,
  Epsilon1 = 0, Epsilon2 = 0), Method = "Mesh", Npool = 0,
  ConvergeTol = 1, Use_REML = FALSE, loc_x = NULL,
  Parameters = "generate", Random = "generate", Map = "generate",
  DiagnosticDir = NULL, TmbDir = system.file("executables", package =
  "VAST"), RunDir = getwd(), build_model = TRUE)
}
\arguments{
\item{TmbData, }{a tagged list of data inputs generated by \code{Data_Fn}}

\item{Version}{a version number (see example for current default).}

\item{RhoConfig}{OPTIONAL, vector of form c("Beta1"=0,"Beta2"=0,"Epsilon1"=0,"Epsilon2"=0) specifying whether either intercepts (Beta1 and Beta2) or spatio-temporal variation (Epsilon1 and Epsilon2) is structured among time intervals (0: each year as fixed effect; 1: each year as random following IID distribution; 2: each year as random following a random walk; 3: constant among years as fixed effect; 4: each year as random following AR1 process)}

\item{Method}{Spatial method used for estimation (determines bounds for logkappa)}

\item{ConvergeTol, }{OPTIONAL override for TMB convergence criteria}

\item{Use_REML, }{OPTIONAL boolean whether to use maximum marginal likelihood or restricted maximum likelihood (termed "REML")}

\item{loc_x}{OPTIONAL, location for each sample used to generate plausible bounds for scale parameter}

\item{Parameters}{OPTIONAL, a tagged list of starting parameters}

\item{Random}{OPTIONAL, a character vector of random effects}

\item{Map}{OPTIONAL, a tagged list of parameters to either mirror or turn off}

\item{DiagnosticDir}{OPTIONAL, a directory where diagonstic runtime information should be stored}

\item{TmbDir}{OPTIONAL, a directory where the CPP file for the VAST model can be found locally}

\item{RunDir}{OPTIONAL, a directory where the CPP file is copied, copiled, and run (must have write privileges or else the function will crash)}
}
\value{
Tagged list containing objects for running a VAST model
\describe{
  \item{Obj}{The built TMB object}
  \item{Upper}{A vector of upper bounds for parameters, optionally for use during optimization}
  \item{Lower}{A vector of lower bounds for parameters, optionally for use during optimization}
  \item{Parameters}{A tagged list of parameter starting values used when building Obj, which can be extracted, modified, and then put back into \code{Build_TMB_Fn} to define different starting values}
  \item{Map}{A taggged list of parameters to be turned off or mirrored, for similar use as Parameters}
  \item{Random}{A character vector of random effects, for similar use as Parameters}
}
}
\description{
\code{Build_TMB_Fn} builds a tagged list with everything necessary to run or interpret inputs to a VAST model
}
