% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{Calc_Anisotropic_Mesh}
\alias{Calc_Anisotropic_Mesh}
\title{Copy of FishStatsUtils::make_mesh}
\usage{
Calc_Anisotropic_Mesh(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{make_mesh}}
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?FishStatsUtils::make_mesh} to see list of arguments
and usage
}
