% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_proportion.R
\name{calculate_proportion}
\alias{calculate_proportion}
\title{Calculate the compositional-expansion}
\usage{
calculate_proportion(TmbData, Index, Year_Set = NULL, Years2Include = NULL,
  strata_names = NULL, category_names = NULL, plot_legend = TRUE,
  DirName = paste0(getwd(), "/"), PlotName = "Proportion.png",
  interval_width = 1, width = 4, height = 4, ...)
}
\arguments{
\item{TmbData}{Formatted data inputs, from `SpatialDeltaGLMM::Data_Fn(...)`}

\item{Index}{output from \code{SpatialDeltaGLMM::PlotIndex_Fn}}

\item{Year_Set}{Year names for labeling panels}

\item{Years2Include}{integer vector, specifying positions of \code{Year_Set} for plotting (used to avoid plotting years with no data, etc.)}

\item{strata_names}{names for spatial strata}

\item{category_names}{names for categories (if using package `VAST`)}

\item{plot_legend}{Add legend for labelling colors}

\item{DirName}{Directory for saving plot and table}

\item{PlotName}{Name for plot}

\item{interval_width}{width for confidence intervals}

\item{width}{plot width in inches}

\item{height}{plot height in inches}

\item{...}{list of settings to pass to \code{sdreport}}
}
\value{
Tagged list of output
\describe{
  \item{Prop_ctl}{Proportion of biomass for each category c, time t, and stratum l}
  \item{Neff_tl}{Effective sample size (median across categories) for each time t and stratum l}
}
}
\description{
\code{calculate_proportion} takes output from a VAST run and calculates the proportion of biomass in different categories
}

