% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fit.R
\name{check_fit}
\alias{check_fit}
\title{Check fit for VAST model}
\usage{
check_fit(parameter_estimates, quiet = FALSE)
}
\arguments{
\item{parameter_estimates}{output from \code{TMBhelper::fit_tmb}}

\item{quiet}{Boolean stating whether to print warnings to terminal}
}
\value{
Return Error code (TRUE is good; FALSE is bad)
}
\description{
\code{check_fit} checks bounds and throws an informative message if any look bad
}
