% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Coherence.R
\name{Coherence}
\alias{Coherence}
\title{Calculate stability metrics}
\usage{
Coherence(Report, Data, covhat = NULL, year_set = NULL)
}
\arguments{
\item{Report}{output report, e.g., from \code{Report <- obj$report()}}

\item{Data}{tagged list of input data}

\item{covhat}{estimated covariance used for calculating coherence}

\item{year_set}{subset of year-indices used when calculating buffering}
}
\value{
Tagged list containing measures of synchrony
\describe{
  \item{phi_x}{Synchrony index for each site}
  \item{phi}{weighted-average of \code{phi_x} weighted by community-abundance at each site}
  \item{psi}{Measure of proportion of variance explained by leading eigen-vectors}
  \item{L_c}{Cholesky decomposition of \code{covhat}}
}
}
\description{
\code{Coherence} calculates buffering (`phi`) and coherence (`psi`)
}

