% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data_Fn.R
\name{Data_Fn}
\alias{Data_Fn}
\title{Build data input for VAST model}
\usage{
Data_Fn(Version, FieldConfig, OverdispersionConfig = c(eta1 = 0, eta2 = 0),
  ObsModel = c(PosDist = 1, Link = 0), b_i, a_i, c_i, s_i, t_i, a_xl,
  MeshList, GridList, Method, v_i = rep(0, length(b_i)), PredTF_i = rep(0,
  length(b_i)), X_xj = NULL, X_xtp = NULL, Q_ik = NULL, Aniso = 1,
  RhoConfig = c(Beta1 = 0, Beta2 = 0, Epsilon1 = 0, Epsilon2 = 0),
  CheckForErrors = TRUE, yearbounds_zz = NULL, Options = c(SD_site_density
  = 0, SD_site_logdensity = 0, Calculate_Range = 0, Calculate_evenness = 0,
  Calculate_effective_area = 0, Calculate_Cov_SE = 0, Calculate_Synchrony = 0,
  Calculate_Coherence = 0))
}
\arguments{
\item{Version}{a version number (see example for current default).}

\item{FieldConfig}{a vector of format c("Omega1"=0, "Epsilon1"=10, "Omega2"="AR1", "Epsilon2"=10), where Omega refers to spatial variation, Epsilon refers to spatio-temporal variation, Omega1 refers to variation in encounter probability, and Omega2 refers to variation in positive catch rates, where 0 is off, "AR1" is an AR1 process, and >0 is the number of elements in a factor-analysis covariance}

\item{OverdispersionConfig}{OPTIONAL, a vector of format c("eta1"=0, "eta2"="AR1") governing any correlated overdispersion among categories for each level of v_i, where eta1 is for encounter probability, and eta2 is for positive catch rates, where 0 is off, "AR1" is an AR1 process, and >0 is the number of elements in a factor-analysis covariance}

\item{ObsModel}{an optimal vector of format c("PosDist"=1,"Link"=0), where PosDist specifies the distribution for positive catch rates (0 is normal, 1 is lognormal, 2 is gamma), and Link is the functional form for encounter probabilities (0 is conventional logit-link, 1 is a novel parameterization involving density)}

\item{b_i}{Sampled biomass for each observation i}

\item{a_i}{Sampled area for each observation i}

\item{c_i}{Category (e.g., species, length-bin) for each observation i}

\item{s_i}{Spatial knot (e.g., grid cell) for each observation i}

\item{t_i}{Time interval (e.g., year) for each observation i}

\item{a_xl}{Area associated with each knot}

\item{MeshList,}{tagged list representing location information for the SPDE mesh hyperdistribution}

\item{GridList,}{tagged list representing location information for the 2D AR1 grid hyperdistribution}

\item{Method,}{character (either "Mesh" or "Grid") specifying hyperdistribution (Default="Mesh")}

\item{v_i}{OPTIONAL, sampling category (e.g., vessel or tow) associated with overdispersed variation for each observation i}

\item{PredTF_i}{OPTIONAL, whether each observation i is included in the likelihood (PredTF_i[i]=0) or in the predictive probability (PredTF_i[i]=1)}

\item{X_xj}{OPTIONAL, matrix of static density covariates (e.g., measured variables affecting density, as used when interpolating density for calculating an index of abundance)}

\item{X_xtp}{OPTIONAL, array of dynamic (varying among time intervals) density covariates}

\item{Q_ik}{OPTIONAL, matrix of catchability covariates (e.g., measured variables affecting catch rates but not caused by variation in species density) for each observation i}

\item{Aniso}{OPTIONAL, whether to assume isotropy (Aniso=0) or geometric anisotropy (Aniso=1)}

\item{RhoConfig}{OPTIONAL, vector of form c("Beta1"=0,"Beta2"=0,"Epsilon1"=0,"Epsilon2"=0) specifying whether either intercepts (Beta1 and Beta2) or spatio-temporal variation (Epsilon1 and Epsilon2) is structured among time intervals (0: each year as fixed effect; 1: each year as random following IID distribution; 2: each year as random following a random walk; 3: constant among years as fixed effect; 4: each year as random following AR1 process)}

\item{CheckForErrors}{OPTIONAL, whether to check for errors in input (NOTE: when CheckForErrors=TRUE, the function will throw an error if it detects a problem with inputs.  However, failing to throw an error is no guaruntee that the inputs are all correct)}

\item{yearbounds_zz}{OPTIONAL, matrix with two columns, giving first and last years for defining one or more periods (rows) used to calculate changes in synchrony over time (only used if \code{Options['Calculate_Synchrony']=1})}

\item{Options}{OPTIONAL, a vector of form c('SD_site_density'=0,'SD_site_logdensity'=0,'Calculate_Range'=0,'Calculate_evenness'=0,'Calculate_effective_area'=0,'Calculate_Cov_SE'=0,'Calculate_Synchrony'=0,'Calculate_Coherence'=0), where Calculate_Range=1 turns on calculation of center of gravity, and Calculate_effective_area=1 turns on calculation of effective area occupied}
}
\value{
Tagged list containing inputs to function Build_TMB_Fn()
}
\description{
\code{Data_Fn} builds a tagged list of data inputs used by TMB for running the model
}

