% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Make_Map}
\alias{Make_Map}
\title{Copy of make_map}
\usage{
Make_Map(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{make_map}
@export}
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?make_map} to see list of arguments and usage
}
