% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plot_factors.R
\name{Plot_factors}
\alias{Plot_factors}
\title{Plot factor-decomposition of covariance}
\usage{
Plot_factors(Report, ParHat, Data, SD, Year_Set = 1:dim(Report$D_xct)[3],
  category_names = 1:dim(Report$D_xct)[2], mapdetails_list = NULL,
  Dim_year = NULL, Dim_species = NULL, plotdir = paste0(getwd(), "/"))
}
\arguments{
\item{Report}{output report, e.g., from \code{Report <- obj$report()}}

\item{ParHat}{Tagged list of fitted data, e.g., from \code{ParHat <- obj$env$parList()}}

\item{Data}{tagged list of input data}

\item{SD}{standard deviation report, e.g., from \code{SD <- sdreport( obj )}}

\item{Year_Set}{plotting-names for time dimension}

\item{category_names}{character-vector listing name for each category}

\item{mapdetails_list}{output from \code{SpatialDeltaGLMM::MapDetails_Fn}}

\item{Dim_year}{Plotting dimension (row,column) for plot of years (default: square with sufficient size for number of years)}

\item{Dim_species}{Plotting dimension (row,column) for plot of categories (default: square with sufficient size for number of categories)}

\item{plotdir}{directory (absolute path) for plots}

\item{Dim_species}{Plotting dimension (row,column) for plot of categories (default: square with sufficient size for number of categories)}
}
\description{
\code{Plot_factors} plots factor loadings, average spatial factors, and spatio-temporal factors
}

