% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_factors.R
\name{Plot_factors}
\alias{Plot_factors}
\title{Plot factor-decomposition of covariance}
\usage{
Plot_factors(Report, ParHat, Data, SD, Year_Set = NULL,
  category_names = NULL, RotationMethod = "PCA",
  mapdetails_list = NULL, Dim_year = NULL, Dim_species = NULL,
  plotdir = paste0(getwd(), "/"), land_color = "grey")
}
\arguments{
\item{Report}{output report, e.g., from \code{Report <- obj$report()}}

\item{ParHat}{Tagged list of fitted data, e.g., from \code{ParHat <- obj$env$parList()}}

\item{Data}{tagged list of input data}

\item{SD}{standard deviation report, e.g., from \code{SD <- sdreport( obj )}}

\item{Year_Set}{plotting-names for time dimension}

\item{category_names}{character-vector listing name for each category}

\item{mapdetails_list}{output from \code{FishStatsUtils::MapDetails_Fn}}

\item{Dim_year}{Plotting dimension (row,column) for plot of years (default: square with sufficient size for number of years)}

\item{Dim_species}{Plotting dimension (row,column) for plot of categories (default: square with sufficient size for number of categories)}

\item{plotdir}{directory for saving plots}

\item{land_color}{color for filling in land (use \code{land_color=rgb(0,0,0,alpha=0)} for transparent land)}
}
\description{
\code{Plot_factors} plots factor loadings, average spatial factors, and spatio-temporal factors
}
