% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{Rerun_Fn}
\alias{Rerun_Fn}
\title{Explore counter-factual scenario}
\usage{
Rerun_Fn(
  parhat0,
  turnoff_pars,
  loc_x,
  cov_to_turnoff = 1:dim(parhat0[["gamma2_ctp"]])[3],
  calculate_COG = TRUE,
  figname = NULL,
  Map = "generate",
  MapDetails_List = NULL,
  year_set = 1:ncol(parhat0[["beta1_ct"]]),
  c_set = 1:nrow(parhat0[["beta1_ct"]]),
  ...
)
}
\arguments{
\item{parhat0}{parameter values to use by default (presumably maximum-likelihood estimates)}

\item{turnoff_pars}{character-vector of parameters to turn off}

\item{loc_x}{location for each knot used to calculate center-of-gravity for counter-factual runs with colnames c('E_km','N_km','Lon','Lat')}

\item{cov_to_turnoff}{which covariates to turn off, as indicated by their order in \code{TmbData$X_xtp} (which only matters if "gamma1_ctp" or "gamma2_ctp" is in \code{turnoff_pars})}

\item{calculate_COG}{Boolean whether to calculate COG for each run}

\item{figname}{name for figure to plot density in counter-factual scenario}

\item{Map}{a tagged list of parameters to either mirror or turn off, using standard TMB interface.
This input is useful, e.g., to build a model without estimating parameters, extracting Map
from the list of outputs, modifying it manually, and then passing it explicitly, \code{make_model(Map=NewMap)} (OPTIONAL)}

\item{MapDetails_List}{output from \code{FishStatsUtils::MapDetails_Fn}}

\item{year_set}{set of parameters to include}

\item{c_set}{set of categories to include}

\item{...}{additional arguments passed to \code{Build_TMB_Fn}}
}
\value{
Tagged list
\describe{
  \item{Report}{Report output for counter-factual run}
  \item{NewBuild_List}{Output from \code{Build_TMB_Fn} using counter-factual parameters}
}
}
\description{
\code{Rerun_Fn} re-builds a model while fixing a subset of parameters at zero
}
\examples{
\dontrun{
# Run without GMRF

Rerun_Fn(parhat0 = Obj$env$parList(),
         turnoff_pars = c("Epsiloninput1_sft", "Epsiloninput2_sft"),
         loc_x = Spatial_List$loc_x,
         TmbData = TmbData, Version = "VAST_v4_0_0")
}
}
