% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAST.R
\docType{package}
\name{VAST}
\alias{VAST}
\alias{VAST-package}
\title{Vector autoregressive spatio-temporal model}
\description{
VAST is multivariate generalization of package `SpatialDeltaGLMM`, and can analyze data for multiple species, ages, size-classes, etc.
}
\details{
See \code{?fit_model} for a simple of example of high-level wrapper functions for using VAST.  Also see examples folder on github \url{http://github.com/james-thorson/VAST/tree/master/examples} for up-to-date examples using lower-level functions.
}
\seealso{
\code{?VAST} for general documentation, \code{?make_settings} for generic settings, \code{?fit_model} for model fitting, and \code{?plot_results} for generic plots
}
